% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whc.R
\name{calc_soil_porosity}
\alias{calc_soil_porosity}
\title{Estimate soil pore space (= theta_sat) (cm3/cm3)}
\usage{
calc_soil_porosity(
  D_BDS,
  A_SAND_MI = NULL,
  A_CLAY_MI = NULL,
  A_SILT_MI = NULL,
  A_SOM_LOI = NULL,
  method = "average"
)
}
\arguments{
\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{method}{(CHAR) The method to estimate soil density. "Heinen" or "average"}
}
\description{
Estimate soil pore space (= theta_sat) (cm3/cm3)
}
