% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_tools.R
\name{check_enum}
\alias{check_enum}
\title{Check enum parameters}
\usage{
check_enum(this.parameter.name, this.parameter.value)
}
\arguments{
\item{this.parameter.name}{(character) Quoted name of the variable to check}

\item{this.parameter.value}{(numeric) Value or vector of values to check}
}
\description{
Check enum parameters
}
