% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_tools.R
\name{check_numeric}
\alias{check_numeric}
\title{Check numeric sptf_parameters}
\usage{
check_numeric(
  this.parameter.name,
  this.parameter.value,
  anymissing = FALSE,
  arg.length = NULL
)
}
\arguments{
\item{this.parameter.name}{(character) Quoted name of the variable to check}

\item{this.parameter.value}{(numeric) Value or vector of values to check}

\item{anymissing}{(boolean) are missing values allowed (TRUE) or not (FALSE)}

\item{arg.length}{(numeric) length of the vector}
}
\description{
Check numeric sptf_parameters
}
