% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmn.R
\name{correct_incubationperiod}
\alias{correct_incubationperiod}
\title{Empirical relationship to convert Nmin to 7 days (mgN/kg/7 days)}
\usage{
correct_incubationperiod(Nmin, t)
}
\arguments{
\item{Nmin}{(numeric) anaerobic N mineralization rate for t-day incubation (mgN/kg)}

\item{t}{(numeric) incubation period (days)}
}
\description{
Empirical relationship to convert Nmin to 7 days (mgN/kg/7 days)
}
