% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_tools.R
\name{enum_opts}
\alias{enum_opts}
\title{Get options for enum parameters}
\usage{
enum_opts(this.parameter)
}
\arguments{
\item{this.parameter}{(character) Quoted name of the parameter}
}
\description{
This function is a wrapper to easily get a vector of options for a parameter of data_type enum
}
\details{
This function wraps unlist(strsplit(sptf_parameters[code == this.parameter, options]))
to return a vector of allowed values for parameters of data_type 'enum'. If the
data is numeric or integer, options will be given as numeric/integer instead of strings
}
\examples{
enum_opts(this.parameter = "B_GWL_CLASS")

}
