% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_tools.R
\name{get_maxval}
\alias{get_maxval}
\title{Get maximum value of a parameter}
\usage{
get_maxval(this.parameter)
}
\arguments{
\item{this.parameter}{(character) Quoted name of the parameter code}
}
\description{
This function retrieves the value highest acceptable value of a numeric parameter
}
\details{
This function retrieves the value in column `value_max` of `sptf_parameters`
Vectors of parameter codes are also possible
}
\examples{
get_maxval(this.parameter = "A_CLAY_MI")

}
