% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_tools.R
\name{get_minval}
\alias{get_minval}
\title{Get minimum value of a parameter}
\usage{
get_minval(this.parameter)
}
\arguments{
\item{this.parameter}{(character) Quoted name of the parameter code}
}
\description{
This function retrieves the value lowest acceptable value of a numeric parameter
}
\details{
This function retrieves the value in column `value_min` of `sptf_parameters`
Vectors of parameter codes are also possible
}
\examples{
get_minval(this.parameter = "A_CLAY_MI")

}
