% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{pr_1}
\alias{pr_1}
\title{Calculate the penetration resistance using ptfs of da Silva and Kay (1997)}
\usage{
pr_1(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
Calculate the penetration resistance using ptfs of da Silva and Kay (1997)
}
\references{
Kay et al. (1997) Sensitivity of soil structure to changes in organic carbon content:Predictions using pedotransfer functions Cites in: da Silva et al. (1997)  Management versus inherent properties effects on bulk density and relative compaction. Soil Tillage Res.
}
