% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptf_predict.R
\name{ptf_bd}
\alias{ptf_bd}
\title{Predict the bulk density with the best combination of existing ptfs from literature.}
\usage{
ptf_bd(
  A_SOM_LOI = NA_real_,
  A_C_OF = NA_real_,
  A_CLAY_MI = NA_real_,
  A_SAND_MI = NA_real_,
  A_SILT_MI = NA_real_,
  B_LU_PTFCLASS = NA_real_,
  B_DEPTH = 0.3,
  B_LOC_COUNTRY = "NL",
  nmax = 5,
  ...
)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage of organic matter in the soil (\%).}

\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}

\item{B_LU_PTFCLASS}{(character) The land use categorie (options: agriculture, grassland, cropland, forest, nature)}

\item{B_DEPTH}{(numeric) The depth of the sampled soil layer (m)}

\item{B_LOC_COUNTRY}{(character) The country code}

\item{nmax}{(integer) the maximum number of ptfs to be included (default nmax = 5)}
}
\description{
Predict the bulk density with the best combination of existing ptfs from literature.
}
\details{
Some of the ptfs require additional information. If given, the relevant ptf's are used, otherwise they are ignored.
These include the total N content (A_N_RT, unit mg/kg), the pH (A_PH_WA), the carbonate content (A_CACO3_MI, unit \%), the moisture content (A_H2O_T105, \%), the slope (B_SLOPE_DEGREE, unit degrees), the aspect (B_SLOPE_ASPECT, unit degrees) and the altidue (B_ALTITUDE, unit m).
When added, please ensure to use the correct element name and units. 

When depth is missing, the function assumes that topsoils are used.
}
