% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptf_predict.R
\name{ptf_bd_all}
\alias{ptf_bd_all}
\title{Predict the bulk density with existing ptfs from literature}
\usage{
ptf_bd_all(dt)
}
\arguments{
\item{dt}{(data table) Data table which includes at least:
A_SOM_LOI or A_C_OF
2 of A_CLAY_MI, A_SAND_MI, A_SILT_MI
B_DEPTH
id}
}
\description{
Predict the bulk density with existing ptfs from literature
}
\details{
This function returns a melted form of data table, containing values of predicted BD with different PTFs
When depth is missing, the function assumes that topsoils are used.
}
