% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilinfo.R
\name{soilinfo.rosetta}
\alias{soilinfo.rosetta}
\title{Get ROSETTA soil parameters from a Texture Class vector}
\usage{
soilinfo.rosetta(textureclass, delog = TRUE, ranges = FALSE)
}
\arguments{
\item{textureclass}{A vector of texture classes (will be sanitized with \link{texture.class})}

\item{delog}{Pass TRUE to de-log the parameters that are log-ed (alpha and n) or ln-ed (Ks and K0)}

\item{ranges}{Pass TRUE to return the min/max of one standard deviation in addition to the values. If
\code{delog=TRUE} these will transformed accordingly, and therefore may not be symmetric around each
value}
}
\value{
A \code{data.frame} with the same number of rows as \code{length(textureclass)}, and the
  following columns:
 \describe{
  \item{theta_r}{Residual water content (cm3/cm3)}
  \item{theta_s}{Saturated water content (cm3/cm3)}
  \item{log_alpha}{The (log10 of the) alpha parameter in the Mualem (1976) model (log 1/cm).
    Will be just \code{alpha} if \code{delog=TRUE}}
  \item{log_n}{The (log10 of the) n parameter in the Mualem (1976) model (log unitless).
    Will be just \code{n} if \code{delog=TRUE}}
  \item{ln_Ks}{The (natural log of the) saturated conductivity (ln cm/day),
    Will be just \code{Ks} if \code{delog=TRUE}}
  \item{ln_K0}{The (natural log of the) K0 conductivity (ln cm/day).
    Will be just \code{Ks} if \code{delog=TRUE}}
  \item{L}{The L parameter in the van Genuchten (1980) model for unsaturated hydraulic conductivity
  (unitless)}
 }

 In addition, columns with the names above and suffixes "_min" and "_max" will be returned
 if \code{ranges=TRUE}.
}
\description{
Get ROSETTA soil parameters from a Texture Class vector
}
\seealso{
rosettaSoilClass
}

