% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd1}
\alias{sptf_bd1}
\title{Calculate the bulk density given the pedotransferfunction in Dutch Fertilizer Recommendation}
\usage{
sptf_bd1(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
This function calculates the bulk density of Dutch soils given texture and organic matter content.
}
\references{
CBAV, 2022; CBGV, 2022
}
