% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd10}
\alias{sptf_bd10}
\title{Calculate the bulk density given the pedotransferfunction of Manrique & Jones (1991).}
\usage{
sptf_bd10(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Manrique & Jones (1991).
}
\references{
Manrique & Jones (1991) Bulk Density of Soils in Relation to Soil Physical and Chemical Properties.
}
