% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd100}
\alias{sptf_bd100}
\title{Calculate the bulk density given the pedotransferfunction of Williams (1970)}
\usage{
sptf_bd100(A_N_RT, A_C_OF)
}
\arguments{
\item{A_N_RT}{(numeric) The nitrogen content of the soil (mg / kg).}

\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Williams (1970)
}
\references{
Williams (1970).Relationships Between the Composition of Soils and Physical Measurements Made on Them
}
