% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd101}
\alias{sptf_bd101}
\title{Calculate the bulk density given the pedotransferfunction of Williams (1970)}
\usage{
sptf_bd101(A_C_OF, A_SAND_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_SAND_MI}{(numeric) Sand content (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Williams (1970)
}
\references{
Williams (1970).Relationships Between the Composition of Soils and Physical Measurements Made on Them
}
