% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd103}
\alias{sptf_bd103}
\title{Calculate the bulk density given the pedotransferfunction of Wu et al. (2003)}
\usage{
sptf_bd103(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Wu et al. (2003)
}
\references{
Wu et al. (2003) Distribution and storage of soil organic carbon in China
}
