% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd104}
\alias{sptf_bd104}
\title{Calculate the bulk density given the pedotransferfunction of Yang et al. (2007)}
\usage{
sptf_bd104(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Yang et al. (2007)
}
\references{
Yang et al. (2007). Storage, patterns and environmental controls of soil organic carbon in China
}
