% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd109}
\alias{sptf_bd109}
\title{Calculate the bulk density given the pedotransferfunction of Zinke et al. (1986).}
\usage{
sptf_bd109(A_C_OF, A_DEPTH)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}

\item{A_DEPTH}{(numeric) The depth of the sampled soil layer (m)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Zinke et al. (1986).
}
\references{
Zinke et al. (1986).Wordlwide organic soil carbon and nitrogen data.
}
