% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd11}
\alias{sptf_bd11}
\title{Calculate the bulk density given the pedotransferfunction of Federer et al. (1993)}
\usage{
sptf_bd11(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Federer et al. (1993)
}
\references{
Federer et al. (1993) The organic fraction - bulk density relationship and the expression of nutrient content in forest soils.
}
