% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd111}
\alias{sptf_bd111}
\title{Calculate the bulk density given the pedotransferfunction of Chen et al. (2018)}
\usage{
sptf_bd111(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Chen et al. (2018)
}
\details{
This pedotransferfunction is an ensemble of six different PTFs. The best one of this study is saved as \code{\link{sptf_bd112}}.
}
\references{
Chen et al. (2018). Building a pedotransfer function for soil bulk density on regional dataset and testing its validity over a larger area
}
