% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd112}
\alias{sptf_bd112}
\title{Calculate the bulk density given the pedotransferfunction of Chen et al. (2018)}
\usage{
sptf_bd112(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Chen et al. (2018)
}
\references{
Chen et al. (2018). Building a pedotransfer function for soil bulk density on regional dataset and testing its validity over a larger area
}
