% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd113}
\alias{sptf_bd113}
\title{Calculate the bulk density given the pedotransferfunction of Souza et al. (2016)}
\usage{
sptf_bd113(A_C_OF, A_CLAY_MI, A_PH_WA)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_PH_WA}{(numeric) Acidity in water (-)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Souza et al. (2016)
}
\references{
Souza et al. (2016) Pedotransfer functions to estimate bulk density from soil properties and environmental covariates: Rio Doce basin
}
