% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd114}
\alias{sptf_bd114}
\title{Calculate the bulk density given the pedotransferfunction of Urbancic et al. (2011)}
\usage{
sptf_bd114(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Urbancic et al. (2011)
}
\references{
Urbancic et al. (2011) PEDOTRANSFER FUNCTIONS FOR BULK DENSITY ESTIMATION OF FOREST SOILS
}
