% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd115}
\alias{sptf_bd115}
\title{Calculate the bulk density given the pedotransferfunction of Hallet et al. (1998)}
\usage{
sptf_bd115(A_C_OF, A_CLAY_MI, A_SAND_MI, A_SILT_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Hallet et al. (1998)
}
\references{
Hallet et al. (1998) Derivation and evaluation of a set of pedogenically-based empirical algorithms for predicting bulk density in British soils
}
