% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd117}
\alias{sptf_bd117}
\title{Calculate the bulk density given the pedotransferfunction of Hossain et al. (2015)}
\usage{
sptf_bd117(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Hossain et al. (2015)
}
\references{
Hossain et al. (2015) Bulk density of mineral and organic soils in the Canada’s arctic and sub-arctic
}
