% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd118}
\alias{sptf_bd118}
\title{Calculate the bulk density given the pedotransferfunction of Hossain et al. (2015)}
\usage{
sptf_bd118(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Hossain et al. (2015)
}
\references{
Hossain et al. (2015) Bulk density of mineral and organic soils in the Canada’s arctic and sub-arctic
}
