% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd119}
\alias{sptf_bd119}
\title{Calculate the bulk density given the pedotransferfunction of Merry et al. (2005)}
\usage{
sptf_bd119(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Merry et al. (2005)
}
\references{
Merry et al. (2005). Used in Sevastas et al. (2018) Predicting bulk density using pedotransfer functions for soils in the Upper Anthemountas basin, Greece
}
