% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd120}
\alias{sptf_bd120}
\title{Calculate the bulk density given the pedotransferfunction of Men et al. (2018).}
\usage{
sptf_bd120(A_C_OF, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Men et al. (2018).
}
\references{
Men et al. (2018) Investigation on Pedotransfer function for estimating soil bulk density in Hebei province
}
