% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd121}
\alias{sptf_bd121}
\title{Calculate the bulk density given the pedotransferfunction of Palladino et al. (2022).}
\usage{
sptf_bd121(
  A_C_OF,
  A_CLAY_MI,
  A_SAND_MI,
  A_CACO3_IF,
  A_PH_WA,
  B_ALTITUDE,
  B_SLOPE_DEGREE
)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}

\item{A_CACO3_IF}{(numeric) The calcium carbonate content of the soil (\%)}

\item{A_PH_WA}{(numeric) The acidity of the soil, pH in water (-)}

\item{B_ALTITUDE}{(numeric) The altitude (m)}

\item{B_SLOPE_DEGREE}{(numeric) The slope of the field (degrees)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Palladino et al. (2022).
}
\references{
Palladino et al. (2022).Developing pedotransfer functions for predicting soil bulk density in Campania
}
