% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd122}
\alias{sptf_bd122}
\title{Calculate the bulk density given the pedotransferfunction of Palladino et al. (2022).}
\usage{
sptf_bd122(A_C_OF, A_CLAY_MI, A_SAND_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Palladino et al. (2022).
}
\references{
Palladino et al. (2022).Developing pedotransfer functions for predicting soil bulk density in Campania
}
