% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd123}
\alias{sptf_bd123}
\title{Calculate the bulk density given the pedotransferfunction of Palladino et al. (2022).}
\usage{
sptf_bd123(A_C_OF, A_CLAY_MI, A_SAND_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SAND_MI}{(numeric) Sand content (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Palladino et al. (2022).
}
\details{
This pedotransferfunction is an ensemble of nine different PTFs, trained for the different regions. 
The best one of this study is saved as \code{\link{sptf_bd122}}.
}
\references{
Palladino et al. (2022).Developing pedotransfer functions for predicting soil bulk density in Campania
}
