% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd124}
\alias{sptf_bd124}
\title{Calculate the bulk density given the pedotransferfunction of Gosselink et al. (1984)}
\usage{
sptf_bd124(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Gosselink et al. (1984)
}
\references{
Gosselink et al. (1984). Relationship of organic carbon and mineral content to bulk density in Louisiana Marsch Soils
}
