% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd13}
\alias{sptf_bd13}
\title{Calculate the bulk density given the pedotransferfunction of Tamminen & Starr (1994)}
\usage{
sptf_bd13(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Tamminen & Starr (1994)
}
\references{
Tamminen & Starr (1994) Bulk density of forested mineral soils.
}
