% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd130}
\alias{sptf_bd130}
\title{Calculate the bulk density given the pedotransferfunction of Foldal et al. (2020).}
\usage{
sptf_bd130(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Foldal et al. (2020).
}
\references{
Foldal et al. (2020). Deriving regional pedotransfer functions to estimate soil bulk density in Austria
}
