% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd135}
\alias{sptf_bd135}
\title{Calculate the bulk density given the pedotransferfunction of Hoogsteen et al. (2020)}
\usage{
sptf_bd135(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Hoogsteen et al. (2020)
}
\references{
Hoogsteen et al. (2020). Monitoring soil organic matter on grassland farms: An exploratory analysis
}
