% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd136}
\alias{sptf_bd136}
\title{Calculate the bulk density given the pedotransferfunction of Khodaverdiloo et al. (2020).}
\usage{
sptf_bd136(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Khodaverdiloo et al. (2020).
}
\references{
Khodaverdiloo et al. (2020). Recalibration of existing pedotransfer functions to estimate soil bulk density at a regional scale
}
