% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd137}
\alias{sptf_bd137}
\title{Calculate the bulk density given the pedotransferfunction of Sun et al. (2019)}
\usage{
sptf_bd137(A_SOM_LOI, A_SILT_MI, A_PH_WA, A_DEPTH)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_PH_WA}{(numeric) Acidity in water (-)}

\item{A_DEPTH}{(numeric) The depth of the sampled soil layer (m)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Sun et al. (2019)
}
\details{
This pedotransferfunction is an ensemble of 6 different PTFs, trained with different soil properties. 
The best one of this study is saved as \code{\link{sptf_bd138}}.
}
\references{
Sun et al. (2019) Comparison of estimated soil bulk density using proximal soil sensing and pedotransfer functions
}
