% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd138}
\alias{sptf_bd138}
\title{Calculate the bulk density given the pedotransferfunction of Sun et al. (2019)}
\usage{
sptf_bd138(A_SOM_LOI, A_SILT_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage of organic matter in the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Sun et al. (2019)
}
\references{
Sun et al. (2019) Comparison of estimated soil bulk density using proximal soil sensing and pedotransfer functions
}
