% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd139}
\alias{sptf_bd139}
\title{Calculate the bulk density given the pedotransferfunction of Pellegrini (2007)}
\usage{
sptf_bd139(A_SOM_LOI, A_CLAY_MI, A_SAND_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage of organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Pellegrini (2007)
}
\references{
Pellegrini (2007). A new pedotransfer function for estimating soil bulk density
}
