% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd14}
\alias{sptf_bd14}
\title{Calculate the bulk density given the pedotransferfunction of Tomasalla & Hodnett (1998).}
\usage{
sptf_bd14(A_C_OF, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Tomasalla & Hodnett (1998).
}
\references{
Tomasalla & Hodnett (1998) Estimating soil water retention characteristics from limited data in Brazilian Amazonia.
}
