% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd143}
\alias{sptf_bd143}
\title{Calculate the bulk density given the pedotransferfunction of Makovnikova et al. (2017).}
\usage{
sptf_bd143(A_C_OF, A_CLAY_MI, A_SAND_MI, A_SILT_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_SILT_MI}{(numeric) Silt content (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Makovnikova et al. (2017).
}
\references{
Makovnikova et al. (2017). Comparison of different models for predicting soil bulk density. Case study – Slovakian agricultural soils
}
