% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd145}
\alias{sptf_bd145}
\title{Calculate the bulk density given the pedotransferfunction of Ramos et al. (2022).}
\usage{
sptf_bd145(A_C_OF, A_CLAY_MI, A_SAND_MI, A_SILT_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The organic C content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Ramos et al. (2022).
}
\details{
This pedotransferfunction is an ensemble of four different PTFs given variable soil types. The best one (general ptf) of this study is saved as \code{\link{sptf_bd144}}.
}
\references{
Ramos et al. (2022).Pedotransference functions for prediction of density in soils of Piauí, Brazil
}
