% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd146}
\alias{sptf_bd146}
\title{Calculate the bulk density given the pedotransferfunction of Katuwal et al. (2020)}
\usage{
sptf_bd146(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Katuwal et al. (2020)
}
\details{
This pedotransferfunction is an ensemble of four different PTFs. The MLR model with other soil properties from this study is saved as \code{\link{sptf_bd147}}.
}
\references{
Katuwal et al. (2020). Predicting the dry bulk density of soils across Denmark: Comparison of single-parameter, multi-parameter, and vis–NIR based models
}
