% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd147}
\alias{sptf_bd147}
\title{Calculate the bulk density given the pedotransferfunction of Katuwal et al. (2020).}
\usage{
sptf_bd147(A_C_OF, A_CLAY_MI, A_SAND_MI, A_SILT_MI, A_DEPTH)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_DEPTH}{(numeric) The depth of the sampled soil layer (m)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Katuwal et al. (2020).
}
\references{
Katuwal et al. (2020). Predicting the dry bulk density of soils across Denmark: Comparison of single-parameter, multi-parameter, and vis–NIR based models
}
