% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd149}
\alias{sptf_bd149}
\title{Calculate the bulk density given the pedotransferfunction of  Chari et al. (2021).}
\usage{
sptf_bd149(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the bulk density given the pedotransferfunction of  Chari et al. (2021).
}
\details{
This pedotransferfunction is an ensemble of four different PTFs. The MLR model with other soil properties from this study is saved as \code{\link{sptf_bd147}}.
}
\references{
Chari et al. (2021). Predicting bulk density using pedotransfer functions for soils in Sistan plain
}
