% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd15}
\alias{sptf_bd15}
\title{Calculate the bulk density given the pedotransferfunction of Bernoux et al. (1998)}
\usage{
sptf_bd15(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Bernoux et al. (1998)
}
\references{
Bernoux et al. (1998) Bulk Densities of Brazilian Amazon Soils Related to Other Soil Properties
}
