% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd150}
\alias{sptf_bd150}
\title{Calculate the bulk density given the pedotransferfunction of Yanti et al. (2021).}
\usage{
sptf_bd150(A_C_OF, A_DEPTH)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}

\item{A_DEPTH}{(numeric) The depth of the sampled soil layer (m)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Yanti et al. (2021).
}
\references{
Yanti et al. (2021). Development of pedotransfer functions for predicting soil bulk density: A case study in Indonesian small island
}
