% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd151}
\alias{sptf_bd151}
\title{Calculate the bulk density given the pedotransferfunction of Sakin et al. (2011)}
\usage{
sptf_bd151(A_CLAY_MI, A_SAND_MI, A_SILT_MI)
}
\arguments{
\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Sakin et al. (2011)
}
\references{
Sakin et al. (2011). Bulk density of Harran plain soils in relation to other soil properties
}
