% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd152}
\alias{sptf_bd152}
\title{Calculate the bulk density given the pedotransferfunction of Sakin et al. (2011).}
\usage{
sptf_bd152(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Sakin et al. (2011).
}
\references{
Sakin et al. (2011). Bulk density of Harran plain soils in relation to other soil properties
}
