% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd153}
\alias{sptf_bd153}
\title{Calculate the bulk density given the pedotransferfunction of Rubinic et al. (2019).}
\usage{
sptf_bd153(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Rubinic et al. (2019).
}
\references{
Rubinic et al. (2019). Prediction of bulk density in Croatian forest Pseudogleys based on contents of soil organic matter and clay
}
