% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd155}
\alias{sptf_bd155}
\title{Calculate the bulk density given the pedotransferfunction of Gomes et al. (2017).}
\usage{
sptf_bd155(A_C_OF, A_SILT_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Gomes et al. (2017).
}
\references{
Gomes et al. (2017). The use of Pedotransfer functions and the estimation of carbon stock in the Central Amazon region
}
