% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd157}
\alias{sptf_bd157}
\title{Calculate the bulk density given the pedotransferfunction of Perreault et al. (2022)}
\usage{
sptf_bd157(A_C_OF, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Perreault et al. (2022)
}
\references{
Perreault et al. (2022) Development of Pedotransfer Functions to Predict Soil Physical Properties in Southern Quebec (Canada)
}
