% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd159}
\alias{sptf_bd159}
\title{Calculate the bulk density given the pedotransferfunction of Hoekstra & Poelman (1982).}
\usage{
sptf_bd159(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Hoekstra & Poelman (1982).
}
\references{
Hoekstra & Poelman (1982). Dichtheid van gronden gemeten aan de meest voorkomende bodemeenheden in Nederland.
}
