% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd16}
\alias{sptf_bd16}
\title{Calculate the bulk density given the pedotransferfunction of Post & Kwon (2000)}
\usage{
sptf_bd16(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Post & Kwon (2000)
}
\references{
Post & Kwon (2000) Soil carbon sequestration and land-use change: processes and potential
}
