% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd160}
\alias{sptf_bd160}
\title{Calculate the bulk density given the pedotransferfunction of Athira et al. (2019).}
\usage{
sptf_bd160(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Athira et al. (2019).
}
\references{
Athira et al. (2019). Influence of soil organic matter on bulk density in Coimbatore soils
}
