% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd161}
\alias{sptf_bd161}
\title{Calculate the bulk density given the pedotransferfunction of Athira et al. (2019).}
\usage{
sptf_bd161(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Athira et al. (2019).
}
\details{
This pedotransferfunction is an ensemble of three different PTFs. The best one of this study is saved as \code{\link{sptf_bd160}}.
}
\references{
Athira et al. (2019). Influence of soil organic matter on bulk density in Coimbatore soils
}
