% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd162}
\alias{sptf_bd162}
\title{Calculate the bulk density given the pedotransferfunction of Sakin et al (2012).}
\usage{
sptf_bd162(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Sakin et al (2012).
}
\references{
Sakin et al (2012). Organic carbon organic matter and bulk density relationships in arid-semi arid soils in Southeast Anatolia region
}
