% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd163}
\alias{sptf_bd163}
\title{Calculate the bulk density given the pedotransferfunction of Tanveera et al (2012).}
\usage{
sptf_bd163(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Tanveera et al (2012).
}
\references{
Tanveera et al (2012). Relation of Soil bulk Density with Texture, Total organic matter content and Porosity in the Soils of Kandi Area of Kashmir valley, India
}
