% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd164}
\alias{sptf_bd164}
\title{Calculate the bulk density given the pedotransferfunction of Shaykewich and Zwarich (1968)}
\usage{
sptf_bd164(A_SOM_LOI, A_CLAY_MI, A_SAND_MI, A_SILT_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage of organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Shaykewich and Zwarich (1968)
}
\references{
Shaykewich and Zwarich (1968). Relationships Between Soil Physical Constants and Soil Physical Components of Some Manitoba Soils
}
