% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd165}
\alias{sptf_bd165}
\title{Calculate the bulk density given the pedotransferfunction of Soane (1975).}
\usage{
sptf_bd165(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Soane (1975).
}
\references{
Soane (1975) Studies on some soil physical properties in relation to cultivations and traffic. Cites in: Soane (1990).
}
