% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd166}
\alias{sptf_bd166}
\title{Calculate the bulk density given the pedotransferfunction of Taulya et al. (2005).}
\usage{
sptf_bd166(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Taulya et al. (2005).
}
\references{
Taulya et al. (2005) Validation of pedotransfer functions for soil bulk density estimation on a Lake Victoria Basin soilscape
}
