% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd167}
\alias{sptf_bd167}
\title{Calculate the bulk density given the pedotransferfunction of Poelman (1975)
note YF: this is the soil density (weight of soil without pore space), not bulk density!}
\usage{
sptf_bd167(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Poelman (1975)
note YF: this is the soil density (weight of soil without pore space), not bulk density!
}
\references{
Poelman (1975) Dichtheid van de vaste delen van rivierkleigronden. In: https://edepot.wur.nl/299664.
}
