% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd169}
\alias{sptf_bd169}
\title{Calculate the bulk density given the pedotransferfunction of Van der Sluijs (1988).}
\usage{
sptf_bd169(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Van der Sluijs (1988).
}
\references{
Van der Sluijs (1988) De poriënindex: een karakteristiek voor vergelijking van de pakking van gronden.
}
