% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd171}
\alias{sptf_bd171}
\title{Calculate the bulk density given the pedotransferfunction of Van der Sluijs (1988).}
\usage{
sptf_bd171(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage of organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Van der Sluijs (1988).
}
\references{
Van der Sluijs (1988) De poriënindex: een karakteristiek voor vergelijking van de pakking van gronden.
}
