% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd172}
\alias{sptf_bd172}
\title{Calculate the bulk density given the pedotransferfunction of Van der Sluijs (1988).}
\usage{
sptf_bd172(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage of organic matter in the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Van der Sluijs (1988).
}
\references{
Van der Sluijs (1988) De poriënindex: een karakteristiek voor vergelijking van de pakking van gronden.
}
