% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd175}
\alias{sptf_bd175}
\title{Calculate the bulk density given the pedotransferfunction of Van Wallenburg (1988).}
\usage{
sptf_bd175(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage of organic matter in the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Van Wallenburg (1988).
}
\references{
Van Wallenburg (1988) De bodemdichtheid van koopveen-, weideveen- en waardveengronden in relatie met bodemkenmerken.
}
