% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd176}
\alias{sptf_bd176}
\title{Calculate the bulk density given the pedotransferfunction of Van Wallenburg (1988).}
\usage{
sptf_bd176(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Van Wallenburg (1988).
}
\references{
Van Wallenburg (1988) De bodemdichtheid van koopveen-, weideveen- en waardveengronden in relatie met bodemkenmerken.
}
