% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd178}
\alias{sptf_bd178}
\title{Calculate the bulk density given the pedotransferfunction of Wosten (1997).}
\usage{
sptf_bd178(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Wosten (1997).
}
\references{
Wosten (1997).Pedotransfer functions to evaluate soil quality.
}
