% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd179}
\alias{sptf_bd179}
\title{Calculate the bulk density given the pedotransferfunction of Melendez (2017).}
\usage{
sptf_bd179(
  A_SOM_LOI,
  A_C_OF = NA,
  A_CLAY_MI = NA,
  A_SILT_MI = NA,
  A_SAND_MI = NA,
  A_PH_WA = NA
)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage of organic matter in the soil (\%).}

\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}

\item{A_PH_WA}{(numeric) The acidity of the soil, pH in water (-)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Melendez (2017).
}
\references{
Melendez (2017) PREDICTING SOIL BULK DENSITY USING HIERARCHICAL PEDOTRANSFER FUNCTIONS AND ARTIFICIAL NEURAL NETWORKS
}
