% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd182}
\alias{sptf_bd182}
\title{Calculate the bulk density given the pedotransferfunction of Ruhlmann et al. (2006)}
\usage{
sptf_bd182(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage of organic matter in the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Ruhlmann et al. (2006)
}
\references{
Ruhlmann et al. (2006). A new approach to calculate the particle density of soils considering properties of the soil organic matter and the mineral matrix. Geoderma 130, 272–283.
}
