% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd183}
\alias{sptf_bd183}
\title{Calculate the bulk density given the pedotransferfunction of McBride et al. (2011)}
\usage{
sptf_bd183(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage of organic matter in the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of McBride et al. (2011)
}
\references{
McBride et al. (2011). Estimating particle density from soil inventory data in the Lake Erie lowlands. Cited in: Schjonning et al. (2016)
}
