% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd184}
\alias{sptf_bd184}
\title{Calculate the bulk density given the pedotransferfunction of Schjonning et al. (2016).}
\usage{
sptf_bd184(A_SOM_LOI, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage of organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Schjonning et al. (2016).
}
\references{
Schjonning et al. (2016) Predicting soil particle density from clay and soil organic matter contents.
}
