% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd187}
\alias{sptf_bd187}
\title{Calculate the bulk density given the pedotransferfunction of Guerif & Farure (1979)}
\usage{
sptf_bd187(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage of organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Guerif & Farure (1979)
}
\references{
Guerif & Farure (1979). Role de la matiere organique sur le comportement des sols au compactage. I. Etude statistique.  Cited in Stengel et al. (1984)
}
