% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd188}
\alias{sptf_bd188}
\title{Calculate the bulk density given the pedotransferfunction of Rawls (1983)}
\usage{
sptf_bd188(A_SOM_LOI, A_CLAY_MI, A_SAND_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage of organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SAND_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Rawls (1983)
}
\references{
Rawls (1983)
}
