% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd189}
\alias{sptf_bd189}
\title{Calculate the bulk density given the pedotransferfunction of Kortleven (1970)}
\usage{
sptf_bd189(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Kortleven (1970)
}
\references{
Kortleven (1970) Volumegewicht, poriënvolume en humusgehalte. Inst. Bodemvruchtbaarheid, Stencil C-7759. Cited in Loman et al. (1972) Kalkverliezen op zandbouwland.
}
