% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd19}
\alias{sptf_bd19}
\title{Calculate the bulk density given the pedotransferfunction of Leonaviciute (2000).}
\usage{
sptf_bd19(A_C_OF, A_CLAY_MI, A_SAND_MI, A_SILT_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Leonaviciute (2000).
}
\references{
Leonaviciute (2000). Predicting soil bulk and particle densities by pedotransfer functions from existing soil data in Lithuania.
}
