% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd190}
\alias{sptf_bd190}
\title{Calculate the bulk density given the pedotransferfunction of Boekel (1961)}
\usage{
sptf_bd190(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Boekel (1961)
}
\references{
Boekel (1961) De bepaling van het soortelijk gewicht van grind. Cited in Schothorst (1963) Het porienvolume van losse en dichte zand- en veengronden. https://edepot.wur.nl/259006
}
