% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd191}
\alias{sptf_bd191}
\title{Calculate the bulk density given the pedotransferfunction of Hooghoudt}
\usage{
sptf_bd191(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Hooghoudt
}
\references{
Hooghoudt (1945) Cited in Schothorst (1963) Het porienvolume van losse en dichte zand- en veengronden. https://edepot.wur.nl/259006
}
