% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd192}
\alias{sptf_bd192}
\title{Calculate the bulk density given the pedotransferfunction of Bryk & Kolodziej (2023)}
\usage{
sptf_bd192(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Bryk & Kolodziej (2023)
}
\references{
Bryk & Kolodziej (2023) Pedotransfer Functions for Estimating Soil Bulk Density Using Image Analysis of Soil Structure
}
