% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd193}
\alias{sptf_bd193}
\title{Calculate the bulk density given the pedotransferfunction of China}
\usage{
sptf_bd193(A_C_OF, A_SAND_MI, A_DEPTH)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_DEPTH}{(numeric) The depth of the sampled soil layer (m)}
}
\description{
Calculate the bulk density given the pedotransferfunction of China
}
\references{
王思楚,王志强,成聪聪,张帅,任霄玉 & 程卓.(2018).东北松嫩黑土区土壤密度传递函数及其适用性. 北京师范大学学报(自然科学版)(03),381-390. doi:10.16360/j.cnki.jbnuns.2018.03.015.
}
