% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd194}
\alias{sptf_bd194}
\title{Calculate the bulk density given the pedotransferfunction in terms of land in China}
\usage{
sptf_bd194(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction in terms of land in China
}
\references{
韩光中,王德彩 & 谢贤健.(2016).中国主要土壤类型的土壤容重传递函数研究. 土壤学报(01),93-102. doi:10.11766/trxb201503300151.
}
