% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd195}
\alias{sptf_bd195}
\title{Calculate the bulk density given the pedotransferfunction for Hebei province in China}
\usage{
sptf_bd195(A_C_OF, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction for Hebei province in China
}
\references{
门明新,彭正萍,许皞 & 宇振荣.(2008).河北省土壤容重的传递函数研究. 土壤通报(01),33-37. doi:10.19336/j.cnki.trtb.2008.01.006.
}
