% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd197}
\alias{sptf_bd197}
\title{Calculate the bulk density given the pedotransferfunction for Heilongjiang province in China}
\usage{
sptf_bd197(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction for Heilongjiang province in China
}
\references{
韩光中 & 李秀芝.(2014).黑龙江省土壤容重传递函数研究. 内江师范学院学报(02),53-55. doi:CNKI:SUN:NJSG.0.2014-02-013.
}
