% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd198}
\alias{sptf_bd198}
\title{Calculate the bulk density given the pedotransferfunction for loess plateau in China}
\usage{
sptf_bd198(A_SAND_MI, A_CLAY_MI)
}
\arguments{
\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction for loess plateau in China
}
\references{
易小波,邵明安,赵春雷 & 张晨成.(2017).黄土高原南北样带不同土层土壤容重变异分析与模拟. 农业机械学报(04),198-205.
}
