% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd199}
\alias{sptf_bd199}
\title{Calculate the bulk density given the pedotransferfunction for Guizhou province in China}
\usage{
sptf_bd199(A_SOM_LOI, A_SAND_MI, A_CLAY_MI, A_N_RT, A_DEPTH)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_N_RT}{(numeric) Total nitrogen content (mg N/ kg)}

\item{A_DEPTH}{(numeric) The depth of the sampled soil layer (m)}
}
\description{
Calculate the bulk density given the pedotransferfunction for Guizhou province in China
}
\references{
李颖 & 周德全.(2017).基于土壤理化性质的贵州省土壤容重传递函数研究. 贵州科学(05),64-71. doi:CNKI:SUN:GZKX.0.2017-05-014.
}
