% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd2}
\alias{sptf_bd2}
\title{Calculate the bulk density given the pedotransferfunction of Curtis & Post (1964).}
\usage{
sptf_bd2(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Curtis & Post (1964).
}
\references{
Curtis & Post (1964) Estimating Bulk Density from Organic-Matter Content in Some Vermont Forest Soils.
}
