% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd200}
\alias{sptf_bd200}
\title{Calculate the bulk density given the pedotransferfunction for Lime soil in China}
\usage{
sptf_bd200(A_SOM_LOI, A_SAND_MI, A_SILT_MI, A_N_RT, A_DEPTH)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_N_RT}{(numeric) Total nitrogen content (mg N/ kg)}

\item{A_DEPTH}{(numeric) The depth of the sampled soil layer (m)}
}
\description{
Calculate the bulk density given the pedotransferfunction for Lime soil in China
}
\references{
李颖,刘秀明,王世杰,周德全 & 罗慧.(2018).中国南方喀斯特地区石灰土容重传递函数模型及影响因素研究. 地球环境学报(03),245-256. doi:CNKI:SUN:DQHJ.0.2018-03-004.
}
