% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd201}
\alias{sptf_bd201}
\title{Calculate the bulk density given the pedotransferfunction for Grassland in Xinjiang province in China}
\usage{
sptf_bd201(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the bulk density given the pedotransferfunction for Grassland in Xinjiang province in China
}
\references{
狄晓双,武红旗,贾宏涛,张文太,谷海斌,邵明轩 & 盛建东.(2021).新疆主要草地土壤容重与有机碳含量关系模型构建. 土壤通报(06),1323-1329. doi:10.19336/j.cnki.trtb.2020101203.
}
