% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd202}
\alias{sptf_bd202}
\title{Calculate the bulk density given the pedotransferfunction for paddy soil in China}
\usage{
sptf_bd202(A_SAND_MI, A_CLAY_MI, A_C_OF)
}
\arguments{
\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the bulk density given the pedotransferfunction for paddy soil in China
}
\references{
Reis, A. M. H. D., Teixeira, W. G., Fontana, A., Barros, A. H. C., De C Victoria, D., Vasques, G. M., Samuel‐Rosa, A., Ottoni, M. V., & De Almeida Monteiro, J. E. B. (2024). Hierarchical pedotransfer functions for predicting bulk density in Brazilian soils. Scientia Agricola, 81. https://doi.org/10.1590/1678-992x-2022-0255
}
