% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd203}
\alias{sptf_bd203}
\title{Calculate the bulk density given the pedotransferfunction for Poland}
\usage{
sptf_bd203(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction for Poland
}
\references{
Bryk, M., & Kołodziej, B. (2023). Pedotransfer functions for estimating soil bulk density using image analysis of soil structure. Sensors, 23(4), 1852. https://doi.org/10.3390/s23041852
}
