% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd204}
\alias{sptf_bd204}
\title{Calculate the bulk density given the pedotransferfunction for U.S. soil}
\usage{
sptf_bd204(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the bulk density given the pedotransferfunction for U.S. soil
}
\references{
Abdelbaki, A. (2018). Evaluation of pedotransfer functions for predicting soil bulk density for U.S. soils. Ain Shams Engineering Journal, 9(4), 1611–1619. https://doi.org/10.1016/j.asej.2016.12.002
}
