% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd22}
\alias{sptf_bd22}
\title{Calculate the bulk density given the pedotransferfunction of Tremblay et al. (2002)}
\usage{
sptf_bd22(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Tremblay et al. (2002)
}
\references{
Tremblay et al. (2002). Prediction of organic carbon content in upland forest soils of Quebec, Canada.
}
