% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd23}
\alias{sptf_bd23}
\title{Calculate the bulk density given the pedotransferfunction of Prevost (2004)}
\usage{
sptf_bd23(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Prevost (2004)
}
\references{
Prevost (2004).Predicting Soil Properties from Organic Matter Content following Mechanical Site Preparation of Forest Soils.
}
