% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd25}
\alias{sptf_bd25}
\title{Calculate the bulk density given the pedotransferfunction of Heuscher et al. (2005)}
\usage{
sptf_bd25(A_C_OF, A_CLAY_MI, A_SILT_MI, A_H2O_T105, A_DEPTH)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_DEPTH}{(numeric) The depth of the sampled soil layer (m)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Heuscher et al. (2005)
}
\references{
Heuscher et al. (2005) Using Soil Physical and Chemical Properties to Estimate Bulk Density.
}
