% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd26}
\alias{sptf_bd26}
\title{Calculate the bulk density given the pedotransferfunction of Perie & Quimet (2007)}
\usage{
sptf_bd26(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Perie & Quimet (2007)
}
\references{
Perie & Quimet (2007). Organic carbon, organic matter and bulk density relationships in boreal forest soils
}
