% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd29}
\alias{sptf_bd29}
\title{Calculate the bulk density given the pedotransferfunction of Tranter et al. (2007)}
\usage{
sptf_bd29(A_C_OF, A_SAND_MI, A_DEPTH)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}

\item{A_DEPTH}{(numeric) The depth of the sampled soil layer (m)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Tranter et al. (2007)
}
\references{
Tranter et al. (2007). Building and testing conceptual and empirical models for predicting soil bulk density
}
