% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd3}
\alias{sptf_bd3}
\title{Calculate the bulk density given the pedotransferfunction of Adams (1973).}
\usage{
sptf_bd3(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Adams (1973).
}
\references{
Adams (1973) THE EFFECT OF ORGANIC MATTER ON THE BULK AND TRUE DENSITIES OF SOME UNCULTIVATED PODZOLIC SOILS.
}
