% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd31}
\alias{sptf_bd31}
\title{Calculate the bulk density given the pedotransferfunction of Benites et al. (2007)}
\usage{
sptf_bd31(A_CLAY_MI)
}
\arguments{
\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Benites et al. (2007)
}
\references{
Benites et al. (2007). Pedotransfer functions for estimating soil bulk density from existing soil survey reports in Brazil.
}
