% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd32}
\alias{sptf_bd32}
\title{Calculate the bulk density given the pedotransferfunction of Ruehlmann & Korschens (2009)}
\usage{
sptf_bd32(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Ruehlmann & Korschens (2009)
}
\references{
Ruehlmann & Korschens (2009) Calculating the Effect of Soil Organic Matter Concentration on Soil Bulk Density
}
