% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd33}
\alias{sptf_bd33}
\title{Calculate the bulk density given the pedotransferfunction of Keller & Hakansson (2010)}
\usage{
sptf_bd33(A_SOM_LOI, A_CLAY_MI, A_SAND_MI, A_SILT_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage of organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Keller & Hakansson (2010)
}
\references{
Keller & Hakansson (2010) Estimation of reference bulk density from soil particle size distribution and soil organic matter content.
}
