% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd35}
\alias{sptf_bd35}
\title{Calculate the bulk density given the pedotransferfunction of Han et al. (2012)}
\usage{
sptf_bd35(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Han et al. (2012)
}
\references{
Han et al. (2012).Pedotransfer Functions for Estimating Soil Bulk Density in China
}
