% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd36}
\alias{sptf_bd36}
\title{Calculate the bulk density given the pedotransferfunction of Hollis et al. (2012)}
\usage{
sptf_bd36(A_C_OF, A_CLAY_MI, A_SAND_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SAND_MI}{(numeric) Sand content (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Hollis et al. (2012)
}
\references{
Hollis et al. (2012). Empirically-derived pedotransfer functions for predicting bulk density in European soils.
}
