% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd37}
\alias{sptf_bd37}
\title{Calculate the bulk density given the pedotransferfunction of Hollis et al. (2012)}
\usage{
sptf_bd37(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Hollis et al. (2012)
}
\references{
Hollis et al. (2012). Empirically-derived pedotransfer functions for predicting bulk density in European soils.
}
