% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd39}
\alias{sptf_bd39}
\title{Calculate the bulk density given the pedotransferfunction of Suuster et al. (2011)}
\usage{
sptf_bd39(A_C_OF, A_CLAY_MI, A_H2O_T105, A_DEPTH)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_H2O_T105}{?soil pore volume? (?)}

\item{A_DEPTH}{(numeric) The depth of the sampled soil layer (m)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Suuster et al. (2011)
}
\references{
Suuster et al. (2011) Soil bulk density pedotransfer functions of the humus horizon in arable soils.
}
