% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd4}
\alias{sptf_bd4}
\title{Calculate the bulk density given the pedotransferfunction of Alexander (1980).}
\usage{
sptf_bd4(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Alexander (1980).
}
\references{
Alexander (1980) Bulk Densities of California Soils in Relation to Other Soil Properties.
}
