% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd40}
\alias{sptf_bd40}
\title{Calculate the bulk density given the pedotransferfunction of Hong et al. (2013)}
\usage{
sptf_bd40(A_SOM_LOI, A_SAND_MI, B_DEPTH)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage of organic matter in the soil (\%).}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}

\item{B_DEPTH}{(numeric) The depth of the sampled soil layer (m)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Hong et al. (2013)
}
\references{
Hong et al. (2013) Predicting and mapping soil available water capacity in Korea.
}
