% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd41}
\alias{sptf_bd41}
\title{Calculate the bulk density given the pedotransferfunction of Hong et al. (2013)}
\usage{
sptf_bd41(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Hong et al. (2013)
}
\references{
Hong et al. (2013) Predicting and mapping soil available water capacity in Korea.
}
