% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd42}
\alias{sptf_bd42}
\title{Calculate the bulk density given the pedotransferfunction of Al-Qinna & Jaber (2013)}
\usage{
sptf_bd42(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Al-Qinna & Jaber (2013)
}
\references{
Al-Qinna & Jaber (2013). PREDICTING SOIL BULK DENSITY USING ADVANCED PEDOTRANSFER FUNCTIONS IN AN ARID ENVIRONMENT
}
