% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd48}
\alias{sptf_bd48}
\title{Calculate the bulk density given the pedotransferfunction of Nanko et al. (2014)}
\usage{
sptf_bd48(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Nanko et al. (2014)
}
\references{
Nanko et al. (2014). A pedotransfer function for estimating bulk density of forest soil in Japan affected by volcanic ash.
}
