% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd49}
\alias{sptf_bd49}
\title{Calculate the bulk density given the pedotransferfunction of Nanko et al. (2014).}
\usage{
sptf_bd49(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Nanko et al. (2014).
}
\references{
Nanko et al. (2014). A pedotransfer function for estimating bulk density of forest soil in Japan affected by volcanic ash.
}
