% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd50}
\alias{sptf_bd50}
\title{Calculate the bulk density given the pedotransferfunction of De Vos et al. (2005).}
\usage{
sptf_bd50(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of De Vos et al. (2005).
}
\references{
De Vos et al. (2005). Predictive Quality of Pedotransfer Functions for Estimating Bulk Density of Forest Soils.
}
