% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd52}
\alias{sptf_bd52}
\title{Calculate the bulk density given the pedotransferfunction of Jeffrey (1970)}
\usage{
sptf_bd52(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Jeffrey (1970)
}
\references{
Jeffrey (1970). A Note on the use of Ignition Loss as a Means for the Approximate Estimation of Soil Bulk Density
}
