% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd57}
\alias{sptf_bd57}
\title{Calculate the bulk density given the pedotransferfunction of Harrison & Bocock (1981)}
\usage{
sptf_bd57(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Harrison & Bocock (1981)
}
\references{
Harrison & Bocock (1981) Estimation of Soil Bulk-Density from Loss-on-Ignition Values.
}
