% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd59}
\alias{sptf_bd59}
\title{Calculate the bulk density given the pedotransferfunction of Grigal (1989)}
\usage{
sptf_bd59(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Grigal (1989)
}
\references{
Grigal (1989). Bulk density of surface soils and peat in the north central united states
}
