% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd6}
\alias{sptf_bd6}
\title{Calculate the bulk density given the pedotransferfunction of Federer (1983).}
\usage{
sptf_bd6(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Federer (1983).
}
\references{
Federer (1983) Nitrogen Mineralization and Nitrification: Depth Variation in Four New England Forest Soils
}
