% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd63}
\alias{sptf_bd63}
\title{Calculate the bulk density given the pedotransferfunction of Honeysett & Ratkowsky (1989).}
\usage{
sptf_bd63(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Honeysett & Ratkowsky (1989).
}
\references{
Honeysett & Ratkowsky (1989). The use of ignition loss to estimate bulk density of forest soils.
}
