% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd64}
\alias{sptf_bd64}
\title{Calculate the bulk density given the pedotransferfunction of Cienciala et al. (2006)}
\usage{
sptf_bd64(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Cienciala et al. (2006)
}
\references{
Cienciala et al. (2006). Forest topsoil organic carbon content in Southwest Bohema region
}
