% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd65}
\alias{sptf_bd65}
\title{Calculate the bulk density given the pedotransferfunction of Aguilera et al. (2013)}
\usage{
sptf_bd65(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Aguilera et al. (2013)
}
\references{
Aguilera et al. (2013). Managing soil carbon for climate change mitigation and adaptation in Mediterranean cropping systems: A meta-analysis
}
