% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd66}
\alias{sptf_bd66}
\title{Calculate the bulk density given the pedotransferfunction of Howard et al. (1995)}
\usage{
sptf_bd66(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Howard et al. (1995)
}
\references{
Howard et al. (1995) The carbon content of soil and its geographical distribution in Great Britain
}
