% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd67}
\alias{sptf_bd67}
\title{Calculate the bulk density given the pedotransferfunction of Song et al. (2005)}
\usage{
sptf_bd67(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Song et al. (2005)
}
\references{
Song et al. (2005) Topsoil organic carbon storage of China and its loss by cultivation
}
