% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd68}
\alias{sptf_bd68}
\title{Calculate the bulk density given the pedotransferfunction of Song et al. (2005)}
\usage{
sptf_bd68(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Song et al. (2005)
}
\references{
Song et al. (2005) Topsoil organic carbon storage of China and its loss by cultivation
}
