% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd69}
\alias{sptf_bd69}
\title{Calculate the bulk density given the pedotransferfunction of Qiao et al. (2019)}
\usage{
sptf_bd69(A_CLAY_MI, A_DEPTH)
}
\arguments{
\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_DEPTH}{(numeric) The depth of the sampled soil layer (m)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Qiao et al. (2019)
}
\references{
Qiao et al. (2019) Development of pedotransfer functions for predicting the bulk density in the critical zone on the Loess Plateau, China
}
