% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd7}
\alias{sptf_bd7}
\title{Calculate the bulk density given the pedotransferfunction of Huntington et al. (1989).}
\usage{
sptf_bd7(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Huntington et al. (1989).
}
\references{
Huntington et al. (1989) Carbon, organic matter and bulk density relationships in a forested spodosol.
}
