% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd70}
\alias{sptf_bd70}
\title{Calculate the bulk density given the pedotransferfunction of Wang et al. (2014)}
\usage{
sptf_bd70(
  A_C_OF,
  A_CLAY_MI,
  A_SILT_MI,
  A_DEPTH,
  B_ALTITUDE,
  B_SLOPE_DEGREE,
  B_SLOPE_ASPECT
)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_DEPTH}{(numeric) The depth of the sampled soil layer (m)}

\item{B_ALTITUDE}{(numeric) The altitude (m)}

\item{B_SLOPE_DEGREE}{(numeric) Slope of the field (degrees)}

\item{B_SLOPE_ASPECT}{(numeric) The slope aspect (degrees)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Wang et al. (2014)
}
\references{
Wang et al. (2014) Prediction of Bulk Density of Soils in the Loess Plateau Region of China
}
