% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd71}
\alias{sptf_bd71}
\title{Calculate the bulk density given the pedotransferfunction of Wang et al. (2014)}
\usage{
sptf_bd71(A_C_OF, A_CLAY_MI, A_SILT_MI, B_SLOPE_DEGREE)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%)}

\item{B_SLOPE_DEGREE}{(numeric) The slope of the field (degrees)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Wang et al. (2014)
}
\references{
Wang et al. (2014) Prediction of Bulk Density of Soils in the Loess Plateau Region of China
}
