% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd74}
\alias{sptf_bd74}
\title{Calculate the bulk density given the pedotransferfunction of Katterer et al (2006)}
\usage{
sptf_bd74(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Katterer et al (2006)
}
\references{
Katterer et al (2006) Pedotransfer functions for estimating plant available water and bulk density in Swedish agricultural soils
}
