% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd79}
\alias{sptf_bd79}
\title{Calculate the bulk density given the pedotransferfunction of Brahim et al. (2012)}
\usage{
sptf_bd79(A_C_OF, A_SILT_MI, A_SAND_MI, A_PH_WA)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_PH_WA}{(numeric) Acidity in water (-)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Brahim et al. (2012)
}
\references{
Brahim et al. (2012). Pedotransfer functions to estimate soil bulk density for Northern Africa: Tunisia case
}
