% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd8}
\alias{sptf_bd8}
\title{Calculate the bulk density given the pedotransferfunction of Huntington et al. (1989).}
\usage{
sptf_bd8(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Huntington et al. (1989).
}
\references{
Huntington et al. (1989) Carbon, organic matter and bulk density relationships in a forested spodosol.
}
