% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd80}
\alias{sptf_bd80}
\title{Calculate the bulk density given the pedotransferfunction of Brahim et al. (2012)}
\usage{
sptf_bd80(A_C_OF, A_CLAY_MI, A_CACO3_IF)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_CACO3_IF}{(numeric) The calcium carbonate content of the soil (\%)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Brahim et al. (2012)
}
\references{
Brahim et al. (2012). Pedotransfer functions to estimate soil bulk density for Northern Africa: Tunisia case
}
