% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd81}
\alias{sptf_bd81}
\title{Calculate the bulk density given the pedotransferfunction of Abeldaki (2016)}
\usage{
sptf_bd81(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Abeldaki (2016)
}
\references{
Abeldaki (2016) Evaluation of pedotransfer functions for predicting soil bulk density for U.S. soils
}
