% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd83}
\alias{sptf_bd83}
\title{Calculate the bulk density given the pedotransferfunction of Apka et al. (2016)}
\usage{
sptf_bd83(A_SAND_MI, A_SILT_MI)
}
\arguments{
\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Apka et al. (2016)
}
\references{
Apka et al. (2016). Enhancing pedotransfer functions with environmental data for estimating bulk density and effective cation exchange capacity in a data-sparse situation
}
