% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd84}
\alias{sptf_bd84}
\title{Calculate the bulk density given the pedotransferfunction of Beutler et al. (2016)}
\usage{
sptf_bd84(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Beutler et al. (2016)
}
\references{
Beutler et al. (2016) Bulk Density Prediction for Histosols and Soil Horizons with High Organic Matter Content
}
