% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd85}
\alias{sptf_bd85}
\title{Calculate the bulk density given the pedotransferfunction of Beutler et al. (2016)}
\usage{
sptf_bd85(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Beutler et al. (2016)
}
\references{
Beutler et al. (2016) Bulk Density Prediction for Histosols and Soil Horizons with High Organic Matter Content
}
