% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd86}
\alias{sptf_bd86}
\title{Calculate the bulk density given the pedotransferfunction of Botula et al. (2015)}
\usage{
sptf_bd86(A_C_OF, A_SAND_MI, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Botula et al. (2015)
}
\references{
Botula et al. (2015). Hierarchical Pedotransfer Functions to Predict Bulk Density of Highly Weathered Soils in Central Africa
}
