% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd88}
\alias{sptf_bd88}
\title{Calculate the bulk density given the pedotransferfunction of Shiri et al. (2017)}
\usage{
sptf_bd88(A_C_OF, A_CLAY_MI, A_CACO3_IF, A_PH_WA)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_CACO3_IF}{(numeric) Carbonated Lime inorganic soil fraction (\%)}

\item{A_PH_WA}{(numeric) Acidity in water (-)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Shiri et al. (2017)
}
\references{
Shiri et al. (2017) Modeling soil bulk density through a complete data scanning procedure: Heuristic alternatives
}
