% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd89}
\alias{sptf_bd89}
\title{Calculate the bulk density given the pedotransferfunction of Calhoun et al. (2001)}
\usage{
sptf_bd89(A_C_OF, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Calhoun et al. (2001)
}
\references{
Calhoun et al. (2001) Predicting bulk density of Ohio Soils from Morphology, Genetic Principles, and Laboratory Characterization Data
}
