% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd91}
\alias{sptf_bd91}
\title{Calculate the bulk density given the pedotransferfunction of Drew (1973)}
\usage{
sptf_bd91(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Drew (1973)
}
\references{
Drew (1973) Bulk Density Estimation Based on Organic Matter Content of Some Minnesota Soils
}
