% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd93}
\alias{sptf_bd93}
\title{Calculate the bulk density given the pedotransferfunction of Kobal et al. (2011)}
\usage{
sptf_bd93(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Kobal et al. (2011)
}
\references{
Kobal et al. (2011) PEDOTRANSFER FUNCTIONS FOR BULK DENSITY ESTIMATION OF FOREST SOILS
}
