% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd94}
\alias{sptf_bd94}
\title{Calculate the bulk density given the pedotransferfunction of Rawls et al. (2004)}
\usage{
sptf_bd94(A_SOM_LOI, A_SAND_MI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage of organic matter in the soil (\%).}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Rawls et al. (2004)
}
\references{
Rawls et al. (2004)
}
