% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd95}
\alias{sptf_bd95}
\title{Calculate the bulk density given the pedotransferfunction of Reidy et al. (2016)}
\usage{
sptf_bd95(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Reidy et al. (2016)
}
\references{
Reidy et al. (2016). Pedotransfer functions for Irish soils – estimation of bulk density (b) per horizon type
}
