% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd96}
\alias{sptf_bd96}
\title{Calculate the bulk density given the pedotransferfunction of Saini (1966)}
\usage{
sptf_bd96(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Saini (1966)
}
\references{
Saini (1966). Organic Matter as a Measure of Bulk Density of Soil
}
