% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd98}
\alias{sptf_bd98}
\title{Calculate the bulk density given the pedotransferfunction of Sevastas et al. (2018)}
\usage{
sptf_bd98(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Sevastas et al. (2018)
}
\references{
Sevastas et al. (2018). Predicting bulk density using pedotransfer functions for soils in the Upper Anthemountas basin, Greece
}
