% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd99}
\alias{sptf_bd99}
\title{Calculate the bulk density given the pedotransferfunction of Williams (1970)}
\usage{
sptf_bd99(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Williams (1970)
}
\references{
Williams (1970).Relationships Between the Composition of Soils and Physical Measurements Made on Them
}
