% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptf_tables.R
\docType{data}
\name{sptf_bouwsteen}
\alias{sptf_bouwsteen}
\title{Table with water retention properties of 'bouwstenen'}
\format{
A data.frame with 36 rows and 14 columns:
\describe{
  \item{bouwsteen}{soil type bouwsteen}
  \item{omschrijving}{description of 'bouwsteen'}
  \item{thres}{residual water content (cm3/cm3). Table 3 of Wosten 2001}
  \item{thsat}{water content at saturation (cm3/cm3). Table 3 of Wosten 2001}
  \item{Ks}{saturated hydraulic conductivity (cm/d). Table 3 of Wosten 2001}
  \item{alpha}{parameter alpha of pF curve (1/cm) Table 3 of Wosten 2001}
  \item{l}{parameter l of pF curve (-). Table 3 of Wosten 2001}
  \item{n}{parameter n of pF curve (-). Table 3 of Wosten 2001}
  \item{sand\%}{sand content (\%) within soil mineral parts. Middle value of Table 1 of Wosten 2001}
  \item{silt\%}{silt content (\%) within soil mineral parts. Middle value of Table 1 of Wosten 2001}
  \item{clay\%}{clay content (\%) within soil mineral parts. Middle value of Table 1 of Wosten 2001}
  \item{OM\%}{organic matter content (\%). Middle value of Table 1 of Wosten 2001}
  \item{bulkdensity}{soil bulk density (g/cm3). Middle value of Table 2 of Wosten 2001}
  \item{M50}{size of sand particles (um). Middle value of Table 2 of Wosten 2001}
}
}
\usage{
sptf_bouwsteen
}
\description{
This table contains water retention curve parameters and typical mineral composition of 18 'bouwstenen'
ref: Wösten, J. H. M., Veerman, G. ., de Groot, W. J., & Stolte, J. (2001). Waterretentie en doorlatendheidskarakteristieken van boven- en ondergronden in Nederland: de Staringreeks. Alterra Rapport, 153, 86. https://edepot.wur.nl/43272
}
\keyword{datasets}
