% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{sptf_cdec1}
\alias{sptf_cdec1}
\title{Calculate the carbon decomposition for peat soils}
\usage{
sptf_cdec1(A_C_OF, A_N_RT, years)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_N_RT}{(numeric) The nitrogen content of the soil (mg / kg).}

\item{years}{(numeric) The years for which the carbon decomposition need to be estimated.}
}
\description{
This function calculates the decomposition of peat using first order kinetics.
}
\references{
Wosten et al. (1997) Bodemkundige vertaalfuncties bij SC-DLO; state of the art. https://edepot.wur.nl/299664
}
