% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec11}
\alias{sptf_cec11}
\title{Calculate the CEC}
\usage{
sptf_cec11(A_SOM_LOI, A_CLAY_MI, A_PH_CC, B_LU_PTFCLASS)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_PH_CC}{(numeric) The acidity of the soil, pH in CaCl2 (-).}

\item{B_LU_PTFCLASS}{(character) The land use category (options: agriculture, grassland, cropland, forest, nature)}
}
\description{
This function calculates the CEC at pH of sample for forest soils (including all horizons) in Switzerland
}
\references{
Soly et al. (2020) A Critical Evaluation of the Relationship Between the Effective Cation Exchange Capacity and Soil Organic Carbon Content in Swiss Forest Soils
}
