% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec12}
\alias{sptf_cec12}
\title{Calculate the CEC}
\usage{
sptf_cec12(A_C_OF, B_LU_PTFCLASS)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{B_LU_PTFCLASS}{(character) The land use category (options: agriculture, grassland, cropland, forest, nature)}
}
\description{
This function calculates the CEC at pH 7 for tropical forest soils in Brazil
}
\references{
Aprile & Lorandi (2012) Evaluation of Cation Exchange Capacity (CEC) in Tropical Soils Using Four Different Analytical Methods
}
