% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec13}
\alias{sptf_cec13}
\title{Calculate the CEC}
\usage{
sptf_cec13(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
This function calculates the CEC at pH 7 of agricultural soils (0-22 cm) in Quebec.
}
\references{
Martel et al. (1978) Cation-exchange capacity of clay-rich soils in relation to organic matter, mineral composition, and surface are.
}
