% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec14}
\alias{sptf_cec14}
\title{Calculate the CEC}
\usage{
sptf_cec14(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
This function calculates the CEC at pH 7 of agricultural and forest top soils (0-22 cm) in Ohio.
}
\references{
wilding & Rutledge (1966) Cation-Exchange capacity as a function of organic matter, total clay, and various clay fractions in a soil toposequence.
}
