% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec15}
\alias{sptf_cec15}
\title{Calculate the CEC}
\usage{
sptf_cec15(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
This function calculates the CEC at pH 8.2 of agricultural soils in Maryland.
}
\references{
Wright & Foss (1972) Contributions of clay and organic matter to the cation exchange capacity of Maryland soils
}
