% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec16}
\alias{sptf_cec16}
\title{Calculate the CEC}
\usage{
sptf_cec16(A_SOM_LOI, A_CLAY_MI, A_PH_WA)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_PH_WA}{(numeric) Acidity in water (-)}
}
\description{
This function calculates the CEC at pH 8.2 for agricultural soils in US
}
\references{
McLean & Owen (1969) Effects of pH on the contributions of organic matter and clay to soil cation exchange capacities.
}
