% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec17}
\alias{sptf_cec17}
\title{Calculate the CEC}
\usage{
sptf_cec17(A_SOM_LOI, B_LU_PTFCLASS)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}

\item{B_LU_PTFCLASS}{(character) The land use category (options: agriculture, grassland, cropland, forest, nature)}
}
\description{
This function calculates the CEC at pH 7 for natural grassland soils in China.
}
\references{
Liu et al. (2020) Soil organic matter and silt contents determine soil particle surface electrochemical properties across a long-term natural restoration grassland
}
