% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec18}
\alias{sptf_cec18}
\title{Calculate the CEC}
\usage{
sptf_cec18(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
This function calculates the CEC at pH 7 of agricultural soils in Florida.
}
\references{
Yuan et al. (1967) Relative contribution of organic and clay fractions to cation-exchange capacity of sandy soils from several soil groups.
}
