% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec19}
\alias{sptf_cec19}
\title{Calculate the CEC}
\usage{
sptf_cec19(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
This function calculates the CEC at pH 8 of agricultural soils in Winconsin.
}
\references{
Helling et al. (1964) Contribution of organic matter and clay to soil cation-exchange capacity as affected by the pH of the saturating solution.
}
