% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec2}
\alias{sptf_cec2}
\title{Calculate the CEC}
\usage{
sptf_cec2(A_SOM_LOI, A_CLAY_MI, A_PH_KCL)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_PH_KCL}{(numeric) Acidity in KCl (-)}
}
\description{
This function calculates the CEC at 8.2
}
\references{
Goselink & van Erp (1999). Cited in NMI-report 655.99.
}
