% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec20}
\alias{sptf_cec20}
\title{Calculate the CEC}
\usage{
sptf_cec20(A_C_OF, A_CLAY_MI, A_PH_WA)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_PH_WA}{(numeric) The acidity of the soil, pH in water (-).}
}
\description{
This function calculates the CEC at pH 7 of forest topsoils (0-30 cm) in Nigeria.
}
\references{
Olorunfemi et al. (2016) Modeling cation exchange capacity and soil water holding capacity from basic soil properties.
}
