% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec21}
\alias{sptf_cec21}
\title{Calculate the CEC}
\usage{
sptf_cec21(A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}
}
\description{
This function calculates the CEC at pH 7 of agricultural soils (variable depth) in Sudan.
}
\references{
Adam et al. (2021) Predicting Soil Cation Exchange Capacity in Entisols with Divergent Textural Classes: The Case of Northern Sudan Soils.
}
