% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec22}
\alias{sptf_cec22}
\title{Calculate the CEC}
\usage{
sptf_cec22(A_SOM_LOI, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}
}
\description{
This function calculates the CEC at pH 8.2 of loess soils (0-20cm) in Argentina
}
\references{
Hepper et al. (2006) Clay mineralogy, cation exchange capacity and specific surface area of loess soils with different volcanic ash contents
}
