% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec23}
\alias{sptf_cec23}
\title{Calculate the CEC}
\usage{
sptf_cec23(A_C_OF, A_CLAY_MI, A_PH_CC)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_PH_CC}{(numeric) The acidity of the soil, pH in CaCl2 (-).}
}
\description{
This function calculates the CEC at pH 7 of forest and nature non-calcareous soils in continental USA.
}
\references{
Seybold et al. (2005) Predicting Cation Exchange Capacity for Soil Survey Using Linear Models
}
