% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec24}
\alias{sptf_cec24}
\title{Calculate the CEC}
\usage{
sptf_cec24(A_C_OF, A_CLAY_MI, A_PH_CC)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_PH_CC}{(numeric) Acidity in CaCl2 (-)}
}
\description{
This function calculates the CEC at pH 7 of forest and nature soils in continental USA.
}
\references{
Seybold et al. (2005) Predicting Cation Exchange Capacity for Soil Survey Using Linear Models
}
