% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec26}
\alias{sptf_cec26}
\title{Calculate the CEC}
\usage{
sptf_cec26(A_C_OF, A_CLAY_MI, A_SILT_MI, A_PH_WA, B_SOILCLASS_USDA)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_PH_WA}{(numeric) Acidity in water (-)}

\item{B_SOILCLASS_USDA}{The soil type class according to the USDA Soil Taxonomy (https://en.wikipedia.org/wiki/USDA_soil_taxonomy)}
}
\description{
This function calculates the CEC at pH 7 of soils in continental USA per taxonomic order stratification group.
}
\references{
Seybold et al. (2005) Predicting Cation Exchange Capacity for Soil Survey Using Linear Models
}
