% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec27}
\alias{sptf_cec27}
\title{Calculate the CEC}
\usage{
sptf_cec27(A_SOM_LOI, A_CLAY_MI, A_SILT_MI, B_CLIM_CAT1)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The clay content of the soil (\%).}

\item{B_CLIM_CAT1}{(character) The climatic region (options: lowland humid (LLH), highland humid (HLH), subhumid (SH) and non humid (NH))}
}
\description{
This function calculates the effective CEC at pH of sample for agricultural topsoils (0-20cm) in Africa
}
\references{
Asadu et al. (1997) A COMPARISON OF THE CONTRIBUTIONS OF CLAY, SILT, AND ORGANIC MATTER TO THE EFFECTIVE CEC OF SOILS OF SUBSAHARAN AFRICA.
}
