% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec28}
\alias{sptf_cec28}
\title{Calculate the CEC}
\usage{
sptf_cec28(A_SOM_LOI, A_CLAY_MI, A_SILT_MI, A_PH_CC)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_PH_CC}{(numeric) Acidity in CaCl2 (-)}
}
\description{
This function calculates the effective CEC at pH of sample for agricultural Danish soils
}
\references{
Krogh et al. (2000) Cation Exchange Capacity Pedotransfer Functions For Danish Soils
}
