% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec29}
\alias{sptf_cec29}
\title{Calculate the CEC}
\usage{
sptf_cec29(A_SOM_LOI, A_CLAY_MI, A_SILT_MI, A_CACO3_IF, A_PH_CC)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}

\item{A_CACO3_IF}{(numeric) The carbonate content of the soil (\%).}

\item{A_PH_CC}{(numeric) The acidity of the soil, pH in CaCl2 (-).}
}
\description{
This function calculates the effective CEC at pH of sample for agricultural Danish topsoils
}
\references{
Krogh et al. (2000) Cation Exchange Capacity Pedotransfer Functions For Danish Soils
}
