% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec3}
\alias{sptf_cec3}
\title{Calculate the CEC}
\usage{
sptf_cec3(A_C_OF, A_CLAY_MI, A_PH_CC)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_PH_CC}{(numeric) Acidity in CaCl2 (-)}
}
\description{
This function calculates the CEC at 8.2
}
\references{
Van Erp et al. (2001) Actual Cation Exchange Capacity of Agricultural soils and its relationship with pH and content of organic carbon and clay
}
