% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec30}
\alias{sptf_cec30}
\title{Calculate the CEC}
\usage{
sptf_cec30(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
This function calculates the effective CEC at sample pH at various depths of agricultural soils in New Jersey
}
\references{
Drake & Motto (1982) An analysis of the effect of clay and organic matter content on the cation exchange capacity of New Jersey Soils.
}
