% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec31}
\alias{sptf_cec31}
\title{Calculate the CEC}
\usage{
sptf_cec31(A_SOM_LOI, A_CLAY_MI, B_SOILCLASS_USDA = NA)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{B_SOILCLASS_USDA}{The soil type class according to the USDA Soil Taxonomy (https://en.wikipedia.org/wiki/USDA_soil_taxonomy)}
}
\description{
This function calculates the effective CEC at sample pH at topsoil of agricultural soils in New Jersey.
}
\details{
Default is the topsoil averaged value. IF the USDA soiltype is known, then the value is updated.
}
\references{
Drake & Motto (1982) An analysis of the effect of clay and organic matter content on the cation exchange capacity of New Jersey Soils.
}
