% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec33}
\alias{sptf_cec33}
\title{Calculate the CEC}
\usage{
sptf_cec33(A_C_OF, A_CLAY_MI, A_PH_CC)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_PH_CC}{(numeric) Acidity in CaCl2 (-)}
}
\description{
This function calculates the CEC at pH 7
}
\details{
Authors also applied ptf from Manrique et al. (1991). Predicting cationexchange capacity from soil physical and chemical properties.
}
\references{
Liao & Zhu (2015) Development of ensemble pedotransfer functions for cation exchange capacity of soils of Qingdao in China.
}
