% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec34}
\alias{sptf_cec34}
\title{Calculate the CEC}
\usage{
sptf_cec34(A_C_OF, A_CLAY_MI, A_PH_CC)

sptf_cec34(A_C_OF, A_CLAY_MI, A_PH_CC)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_PH_CC}{(numeric) The acidity of the soil, pH in CaCl2 (-)}
}
\description{
This function calculates the CEC at pH 7

This function calculates the CEC at pH 7
}
\references{
Thiere et al. (1986). Analyse des Zusammenhanges von Austauschkapazit€at (T-Wert), Kornung, organischer Bodensubstanz und Bodenreaktion fur verbreitete € Substrat- und Horizontgruppen. Cited in Liao & Zhu (2015).

Meyer et al. (1994).  Cation exchange capacities of upland soils in eastern Canada. Cited in Liao & Zhu (2015).
}
