% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec35}
\alias{sptf_cec35}
\title{Calculate the CEC}
\usage{
sptf_cec35(A_C_OF, A_CLAY_MI, A_PH_CC)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_PH_CC}{(numeric) The acidity of the soil, pH in CaCl2 (-)}
}
\description{
This function calculates the CEC at pH 7
}
\references{
Scheinost (1995). Pedotransfer-Funktionen zum Wasser-und Stoffhaushalt einer Bodenlandschaft. FAM-Bericht 5, Shaker, Aachen. Cited in Liao & Zhu (2015).
}
