% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec36}
\alias{sptf_cec36}
\title{Calculate the CEC}
\usage{
sptf_cec36(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
This function calculates the CEC at pH 7
}
\references{
Foth & Ellis (1997). Soil fertility, 2nd edn. Lewis CRC Press LLC, Boston, USA. pp. 57. Cited in Liao & Zhu (2015).
}
