% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec37}
\alias{sptf_cec37}
\title{Calculate the CEC}
\usage{
sptf_cec37(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
This function calculates the CEC at pH 8.2 in agricultural soils in Indiana.
}
\references{
Franzmeier et al. (1990). Relation of cation exchange capacity to clay and organic carbon contents of Indiana soils.
}
