% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec38}
\alias{sptf_cec38}
\title{Calculate the CEC}
\usage{
sptf_cec38(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
This function calculates the CEC at pH 7 for chernozemic soils in Saskatchewan province of Canada.
}
\references{
Arnaud & Sephton (1972). Contribution of clay and organic matter to cation-exchange capacity of chemozemic soils.
}
