% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec39}
\alias{sptf_cec39}
\title{Calculate the CEC}
\usage{
sptf_cec39(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
This function calculates the CEC at pH 7
}
\references{
Brady (1990). The nature and properties of soils, 1Oth edn. Macmillan Co, New York, NY. Cited in Liao & Zhu (2015).
}
