% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec4}
\alias{sptf_cec4}
\title{Calculate the CEC}
\usage{
sptf_cec4(A_C_OF, A_CLAY_MI, A_PH_CC)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_PH_CC}{(numeric) Acidity in CaCl2 (-)}
}
\description{
This function calculates the CEC at pH 6.5 for agricultural and natural soils in Winconsin (USA).
}
\references{
Helling et al. (1964) Contribution of organic matter and clay to soil cation exchange capacity as affected by the pH of the saturating soil solution.
}
