% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec40}
\alias{sptf_cec40}
\title{Calculate the CEC}
\usage{
sptf_cec40(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
This function calculates the CEC at pH 7
}
\references{
MacDonald (1998). Development of pedotransfer functions of southern Ontario Soils. Cited in Liao & Zhu (2015).
}
