% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec41}
\alias{sptf_cec41}
\title{Calculate the CEC}
\usage{
sptf_cec41(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
This function calculates the CEC at pH 7 for forests soils in Mississippi, Ohio, Wisconsin, North Carolina and Florida.
}
\references{
Miller (1970). Inter-regional predictability of cation-exchange capacity by multiple regression.
}
