% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec42}
\alias{sptf_cec42}
\title{Calculate the CEC}
\usage{
sptf_cec42(A_C_OF, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}
}
\description{
This function calculates the CEC at pH 7 for agricultural soils in Cameroon
}
\references{
Enang et al. (2022) Pedotransfer functions for cation exchange capacity estimation in highly weathered soils of the tropical highlands of NW Cameroon
}
