% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec44}
\alias{sptf_cec44}
\title{Calculate the CEC}
\usage{
sptf_cec44(A_C_OF, A_CLAY_MI, A_SILT_MI, A_PH_CC)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_PH_CC}{(numeric) Acidity in CaCl2 (-)}
}
\description{
This function calculates the effective CEC at sample pH for agricultural topsoils in Nigeria
}
\references{
Apka et al. (2016) Enhancing pedotransfer functions with environmental data for estimating bulk density and effective cation exchange capacity in a data-sparse situation
}
