% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec45}
\alias{sptf_cec45}
\title{Calculate the CEC}
\usage{
sptf_cec45(A_C_OF, A_CLAY_MI, A_PH_WA)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_PH_WA}{(numeric) The acidity of the soil, pH in water (-)}
}
\description{
This function calculates the CEC at pH 7 at agricultural sites in Finland
}
\references{
Raty et al. (2021) Estimating cation exchange capacity and clay content from agricultural soil testing data.
}
