% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec47}
\alias{sptf_cec47}
\title{Calculate the CEC}
\usage{
sptf_cec47(A_SOM_LOI, A_CLAY_MI, A_SILT_MI, B_LU_PTFCLASS)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{B_LU_PTFCLASS}{(character) The land use category (options: agriculture, grassland, cropland, forest, nature)}
}
\description{
This function calculates the effective CEC at pH of sample for agricultural topsoils (0-20cm) in Africa
}
\references{
Asadu et al. (1997) Contributions of organic matter, clay and silt to the effective CEC of soils of different land use history
}
