% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec48}
\alias{sptf_cec48}
\title{Calculate the CEC}
\usage{
sptf_cec48(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
This function calculates the CEC at pH 6.5 for agricultural fields at two depths in Central Iran.
}
\references{
Salehi et al. (2008) Developing Soil Cation Exchange Capacity Pedotransfer Functions using Regression and Neural Networks and the Effect of Soil Partitioning on the Accuracy and Precision of Estimation
}
