% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec49}
\alias{sptf_cec49}
\title{Calculate the CEC}
\usage{
sptf_cec49(A_SOM_LOI, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}
}
\description{
This function calculates the effective CEC at pH 7 for agricultural topsoils (0-30cm) in Iraqi
}
\references{
Fattah & Karim (2021) PERFORMANCE OF LINEAR MODELS IN PREDICTING CATION EXCHANGE CAPACITY OF CALCAREOUS SOILS
}
