% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec5}
\alias{sptf_cec5}
\title{Calculate the CEC}
\usage{
sptf_cec5(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
This function calculates the CEC at pH 8.2 for agricultural soils in the Netherlands.
}
\references{
Bril (unknown). Cited in Bussink & van de Zande (2008) Achtergrondinformatie over CEC en CEC-inbedding in bemestingsadviezen
}
