% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec52}
\alias{sptf_cec52}
\title{Calculate the CEC}
\usage{
sptf_cec52(A_C_OF, A_CLAY_MI, A_SILT_MI, A_PH_WA)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_PH_WA}{(numeric) Acidity in water (-)}
}
\description{
This function calculates the CEC at pH 8.2 for various land uses in Iraq
}
\references{
Khaledian et al. (2017). Modeling soil cation exchange capacity in multiple countries.
}
