% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec53}
\alias{sptf_cec53}
\title{Calculate the CEC}
\usage{
sptf_cec53(A_C_OF, A_CLAY_MI, A_SILT_MI, A_PH_WA)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}

\item{A_PH_WA}{(numeric) The acidity of the soil, pH in water (-).}
}
\description{
This function calculates the CEC at pH 8.2 for various land uses in Spain
}
\references{
Khaledian et al. (2017). Modeling soil cation exchange capacity in multiple countries.
}
