% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec54}
\alias{sptf_cec54}
\title{Calculate the CEC}
\usage{
sptf_cec54(A_C_OF, A_CLAY_MI, A_SILT_MI, A_PH_WA, B_LU_PTFCLASS)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_PH_WA}{(numeric) Acidity in water (-)}

\item{B_LU_PTFCLASS}{(character) The land use category (options: agriculture, grassland, cropland, forest, nature)}
}
\description{
This function calculates the CEC at pH 8.2 for various land uses in USA, Spain, Iran and Iraq.
}
\references{
Khaledian et al. (2017). Modeling soil cation exchange capacity in multiple countries.
}
