% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec56}
\alias{sptf_cec56}
\title{Calculate the CEC}
\usage{
sptf_cec56(A_SOM_LOI, A_CLAY_MI, A_PH_WA)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_PH_WA}{(numeric) Acidity in water (-)}
}
\description{
This function calculates the CEC at pH 8.2 for calcareous soils (0-20cm) in Mexico
}
\references{
Bell & van Keulen (1995). Soil pedotransfer functions for 4 Mexican soils. Cited in Razzaghi et al. (2021).
}
