% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec57}
\alias{sptf_cec57}
\title{Calculate the CEC}
\usage{
sptf_cec57(A_SOM_LOI, A_SAND_MI, A_PH_WA)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_PH_WA}{(numeric) Acidity in water (-)}
}
\description{
This function calculates the CEC at pH 8.2 for agricultural soils (0-20cm) in Mexico
}
\references{
Bell (1993). Organic matter, soil properties and wheat production in the high Valley of Mexico
}
