% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec58}
\alias{sptf_cec58}
\title{Calculate the CEC}
\usage{
sptf_cec58(A_SOM_LOI, A_CLAY_MI, A_SILT_MI, A_CACO3_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_CACO3_MI}{(numeric) The calcium carbonate content of the soil (\%)}
}
\description{
This function calculates the CEC at pH 8.2 for calcareous soils in Iran
}
\references{
Asadzadeh et al. (2019). Predicting cationic exhcnage capacity in calcareous soils of East Azerbaijen province, northwest Iran. Cited in Razzaghi et al. (2021).
}
