% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec6}
\alias{sptf_cec6}
\title{Calculate the CEC}
\usage{
sptf_cec6(A_SOM_LOI, A_CLAY_MI, A_SILT_MI, A_PH_CC)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%)}

\item{A_PH_CC}{(numeric) The acidity of the soil, pH in CaCl2 (-)}
}
\description{
This function calculates the CEC at pH 8.2 for agricultural soils in the Netherlands.
}
\references{
Houba (unknown). Cited in Bussink & van de Zande (2008) Achtergrondinformatie over CEC en CEC-inbedding in bemestingsadviezen
}
