% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec60}
\alias{sptf_cec60}
\title{Calculate the CEC}
\usage{
sptf_cec60(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
This function calculates the CEC at pH 8.2 for calcareous soils in various countries across all continents
}
\references{
Rehman et al. (2019). Comparison of cation exchange capacity estimated from Vis-NIR spectral relfectance data and a pedotransfer function. Cited in Razzaghi et al. (2021).
}
