% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec61}
\alias{sptf_cec61}
\title{Calculate the CEC}
\usage{
sptf_cec61(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
This function calculates the CEC at pH 8.2 for calcareous soils in Iran
}
\references{
Khodaverdiloo et al. (2018). Performance of soil cation exchange capacity pedotransfer function as affected by the inputs and database size. Cited in Razzaghi et al. (2021).
}
