% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec64}
\alias{sptf_cec64}
\title{Calculate the CEC}
\usage{
sptf_cec64(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
This function calculates the CEC for agricultural soils (0-20cm) in New Zealand
}
\references{
Curtin & Trolove (2013) Predicting pH buffering capacity of New Zealand soils from organic matter content and mineral characteristics
}
