% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec65}
\alias{sptf_cec65}
\title{Calculate the CEC}
\usage{
sptf_cec65(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
This function calculates the effective CEC at pH 7 for agricultural topsoils (0-30cm) in Iran.
}
\references{
Rashidi & Seilsepour (2008) Modeling of soil cation exchange capacity based on soil organic carbon
}
