% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec66}
\alias{sptf_cec66}
\title{Calculate the CEC}
\usage{
sptf_cec66(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
This function calculates the CEC for soils in Iraq
}
\references{
Majed et al. (1991) Relative contribution of clay and organic fraction to the cation exchange capacity of northern Iraqi Soils. Cited in Fattah et al. (2021)
}
