% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec67}
\alias{sptf_cec67}
\title{Calculate the CEC}
\usage{
sptf_cec67(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
This function calculates the CEC for soils in Iran
}
\references{
Sarmadian et al. (2009) Modeling of Some Soil Properties Using Artificial Neural Network and Multivariate Regression in Gorgan Province, North Iran
}
