% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec68}
\alias{sptf_cec68}
\title{Calculate the CEC}
\usage{
sptf_cec68(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
This function calculates the CEC for soils in New Zealand
}
\references{
Parfitt et al. (1995) Contribution of organic matter and clay minerals to the cation exchange capacity of soils
}
