% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec69}
\alias{sptf_cec69}
\title{Calculate the CEC}
\usage{
sptf_cec69(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
This function calculates the CEC at pH 7 for soils in India, variable land uses
}
\references{
Somani (1983). The computation of the cation exchange capacity of soils from clay and organic matter content. Cited in Meyer (1994).
}
