% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec70}
\alias{sptf_cec70}
\title{Calculate the CEC}
\usage{
sptf_cec70(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
This function calculates the CEC at pH 7 for forage soils in Quebec.
}
\references{
Martel and Laverdiere (1976). Facteurs qui influencent la teneur de la matibre organique et les propietes d'echange cationique des horizons Ap des sols de grande culture du Quebec. Cited in Meyer (1994).
}
