% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec71}
\alias{sptf_cec71}
\title{Calculate the CEC}
\usage{
sptf_cec71(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
This function calculates the CEC at pH 8.2 for various land use soils in Ireland
}
\references{
Curtin and Smillie (1976). Estimation of components of soil cation exchange capacity from measurements of specific surface and organic matter. Cited in Meyer (1994).
}
