% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec72}
\alias{sptf_cec72}
\title{Calculate the CEC}
\usage{
sptf_cec72(A_C_OF, A_CLAY_MI, A_PH_WA)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_PH_WA}{(numeric) Acidity in water (-)}
}
\description{
This function calculates the CEC at pH 8.2 for various land use soils in Ireland
}
\references{
Tranter et al. (2009). Using distance metrics to determine the appropriate domain of pedotransfer function predictions
}
