% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec73}
\alias{sptf_cec73}
\title{Calculate the CEC}
\usage{
sptf_cec73(A_SOM_LOI, A_CLAY_MI, A_SILT_MI, A_PH_WA)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The clay content of the soil (\%).}

\item{A_PH_WA}{(numeric) The acidity of the soil, pH in water (-).}
}
\description{
This function calculates the effective CEC at pH 7 for agricultural topsoils (0-30cm) in Germany
}
\references{
deumlich et al. (2015) Characterization of cation exchange capacity (CEC) for agricultural land-use areas
}
