% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec74}
\alias{sptf_cec74}
\title{Calculate the CEC}
\usage{
sptf_cec74(A_C_OF, A_CLAY_MI, A_SILT_MI, A_PH_WA, B_SOILCLASS_USDA)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_PH_WA}{(numeric) Acidity in water (-)}

\item{B_SOILCLASS_USDA}{The soil type class according to the USDA Soil Taxonomy (https://en.wikipedia.org/wiki/USDA_soil_taxonomy)}
}
\description{
This function calculates the CEC at pH 7 of soils in continental USA per taxonomic order stratification group.
}
\references{
Manrique et al. (1990) Estimation of exchangeable bases and base saturation from soil physical and chemical data
}
