% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec76}
\alias{sptf_cec76}
\title{Calculate the CEC}
\usage{
sptf_cec76(A_SOM_LOI, A_SAND_MI, A_PH_WA)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}

\item{A_PH_WA}{(numeric) The acidity of the soil, pH in water (-)}
}
\description{
This function calculates the CEC at pH 7 from agricultural topsoils (0-20 cm) in China
}
\references{
Yunan et al. (2018) Study on Cation Exchange Capacity of Agricultural Soils
}
