% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec77}
\alias{sptf_cec77}
\title{Calculate the CEC}
\usage{
sptf_cec77(A_C_OF, A_CLAY_MI, A_SILT_MI, A_SAND_MI, A_PH_WA)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}

\item{A_PH_WA}{(numeric) The acidity of the soil, pH in water (-)}
}
\description{
This function calculates the CEC at pH 7 from agricultural topsoils (0-30 cm) in Iran
}
\references{
Ghorbani et al. (2015) Estimation of Soil Cation Exchange Capacity using Multiple Regression, Artificial Neural Networks, and Adaptive Neuro-fuzzy Inference System Models in Golestan Province, Iran
}
