% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec78}
\alias{sptf_cec78}
\title{Calculate the CEC}
\usage{
sptf_cec78(A_SOM_LOI, A_CLAY_MI, A_SILT_MI, A_PH_WA)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}

\item{A_PH_WA}{(numeric) The acidity of the soil, pH in water (-)}
}
\description{
This function calculates the CEC at pH 7 from agricultural topsoils (0-30 cm) in China
}
\references{
Wang et al. (2012) Distribution and Affecting Factors of Soil Cation Exchange Capacity in Watershed of the Loess Plateau. Cited in Yunan et al. (2018).
}
