% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec79}
\alias{sptf_cec79}
\title{Calculate the CEC}
\usage{
sptf_cec79(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
This function calculates the CEC at pH 8.2 for various land use soils in Australia
}
\references{
Hallsworth and Wilkinson (1958). The contribution of clay and orgnai cmatter to the cation exchange capacity of soils.
}
