% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec8}
\alias{sptf_cec8}
\title{Calculate the CEC}
\usage{
sptf_cec8(A_SOM_LOI, A_CLAY_MI, A_PH_CC, A_CN_FR, B_LU_PTFCLASS)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_PH_CC}{(numeric) The acidity of the soil, pH in CaCl2 (-).}

\item{A_CN_FR}{(numeric) The carbon-to-nitrogen ratio of the soil organic matter (-)}

\item{B_LU_PTFCLASS}{(character) The land use category (options: agriculture, grassland, cropland, forest, nature)}
}
\description{
This function calculates the CEC at pH 8.2 for peaty agricultural soils in the Netherlands.
}
\references{
Ros & Reijneveld (2015) Bodemadvies voor gips en kalkmeststoffen. NMI-report 1588.N.15.
}
