% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec80}
\alias{sptf_cec80}
\title{Calculate the CEC}
\usage{
sptf_cec80(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}
}
\description{
This function calculates the CEC at pH 8.2 for agricultural red soil in Australia
}
\references{
Chan et al. (1992) Organic Carbon and Associated Soil Properties of a Red Earth after 10 years of Rotation under Different Stubble and Tillage Practices
}
