% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec81}
\alias{sptf_cec81}
\title{Calculate the CEC}
\usage{
sptf_cec81(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}
}
\description{
This function calculates the CEC at pH 8.2 for agricultural red soil in Spain
}
\references{
Caravaca et al. (1991). Organic matter, nutrient contents and cation exchange capacity in fine fractions from semiarid calcareous soils
}
