% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec9}
\alias{sptf_cec9}
\title{Calculate the CEC}
\usage{
sptf_cec9(A_SOM_LOI, A_CLAY_MI, A_PH_CC, B_LU_PTFCLASS)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_PH_CC}{(numeric) Acidity in CaCl2 (-)}

\item{B_LU_PTFCLASS}{(character) The land use category (options: agriculture, grassland, cropland, forest, nature)}
}
\description{
This function calculates the CEC at pH 8.2 for clay agricultural soils
}
\references{
Ros & Reijneveld (2015) Bodemadvies voor gips en kalkmeststoffen. NMI-report 1588.N.15.
}
