% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erodibility.R
\name{sptf_ef4}
\alias{sptf_ef4}
\title{Calculate the wind erodible fraction (-)}
\usage{
sptf_ef4(A_CLAY_MI, A_SAND_MI, A_SOM_LOI)
}
\arguments{
\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
This function calculates the erodible fraction in Argentina(main) and Spain
}
\references{
reference López, M., De Dios Herrero, J. M., Hevia, G., Gracia, R., & Buschiazzo, D. E. (2007). Determination of the wind-erodible fraction of soils using different methodologies. Geoderma, 139(3–4), 407–411. https://doi.org/10.1016/j.geoderma.2007.03.006
}
