% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erodibility.R
\name{sptf_ef5}
\alias{sptf_ef5}
\title{Calculate the wind erodible fraction (-)}
\usage{
sptf_ef5(A_CLAY_MI, A_SAND_MI, A_SILT_MI, A_SOM_LOI, A_CACO3_IF)
}
\arguments{
\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CACO3_IF}{(numeric) Carbonated Lime inorganic soil fraction (\%)}
}
\description{
This function calculates the erodible fraction in U.S.
}
\references{
Fryrear, D. W., Krammes, C. A., Williamson, D. L., & Zobeck, T. M. (1994). Computing the wind erodible fraction of soils. Journal of Soil and Water Conservation, 49(2), 183-188.
}
