% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erodibility.R
\name{sptf_ef7}
\alias{sptf_ef7}
\title{Calculate the Erodible fraction(EF)}
\usage{
sptf_ef7(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
Calculate the erodible fraction for Kassala State in Sudan
}
\references{
Hassan, A. A., & Mustafa, M. A. (2011). Assessment and mapping of wind erodibility of Aridisols and Entisols in the river Nile State, Sudan. In The 5th Annual Conference-Agricultural and Veterinary Research-February 2014. University of Khartoum
}
