% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erodibility.R
\name{sptf_erodibility1}
\alias{sptf_erodibility1}
\title{Calculate the wind erodibility as a function of MWD.}
\usage{
sptf_erodibility1(A_AMWD_DM)
}
\arguments{
\item{A_AMWD_DM}{(numeric) The aggregate mean weight diameter in mm (dry aggregate stability).}
}
\description{
This function calculates the wind erodibility as a function of MWD for soils (0-20cm) in Iran, being the slope of wind erosivity plotted to erosion rate.
}
\references{
Zamani & Mahmoodabadi (2013) Effect of particle-size distribution on wind erosion rate and soil erodibility
}
