% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erodibility.R
\name{sptf_erodibility2}
\alias{sptf_erodibility2}
\title{Calculate the wind erodible fraction (-)}
\usage{
sptf_erodibility2(A_CLAY_MI, A_SILT_MI, A_C_OF)
}
\arguments{
\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
This function calculates the erodible fraction in Argentina
}
\references{
Colazo & Buschiazzo (2010) Soil dry aggregate stability and wind erodible fraction in a semiarid environment of Argentina
}
