% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metals.R
\name{sptf_fc_cu}
\alias{sptf_fc_cu}
\title{Calculate the copper freundlich coefficient determining the adsorption rate of copper in soils}
\usage{
sptf_fc_cu(A_SOM_LOI, A_CLAY_MI, A_PH_WA)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_PH_WA}{(numeric) Acidity in water (-)}
}
\description{
This function calculates the freundlich coefficient (mol/ l n /kg) for cu from soil properties
}
\references{
Romkens et al. (2004) Derivation of parition relationships to calculate Cd, Cu, Ni, Pb and Zn solubility and activity in soil solutions.
}
