% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metals.R
\name{sptf_fc_zn}
\alias{sptf_fc_zn}
\title{Calculate the zinc freundlich coefficient determining the adsorption rate of zinc in soils}
\usage{
sptf_fc_zn(A_SOM_LOI, A_CLAY_MI, A_PH_WA)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_PH_WA}{(numeric) The acidity of the soil, pH in water (-)}
}
\description{
This function calculates the freundlich coefficient (mol/ l n /kg) for Zn from soil properties
}
\references{
Romkens et al. (2004) Derivation of parition relationships to calculate Cd, Cu, Ni, Pb and Zn solubility and activity in soil solutions.
}
