% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwc.R
\name{sptf_hwc1}
\alias{sptf_hwc1}
\title{Calculate the HWC content}
\usage{
sptf_hwc1(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
This function calculates the hot water extractable carbon content for agricultural soils in the Netherlands.
}
\references{
Ros (2011) Predicting soil Nitrogen Supply.
}
