% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwc.R
\name{sptf_hwc10}
\alias{sptf_hwc10}
\title{Calculate the HWC content}
\usage{
sptf_hwc10(A_C_OF, A_PH_CC)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_PH_CC}{(numeric) The acidity of the soil, pH in CaCl2 (-)}
}
\description{
This function calculates the hot water extractable carbon content for agricultural soils in the Netherlands. Data from various years.
}
\references{
Hanegraaf et al. (2009). De afbraaksnelheid van organische stof in Drenthe. NMI project 972.N.09, 93 pp.
}
