% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwc.R
\name{sptf_hwc12}
\alias{sptf_hwc12}
\title{Calculate the HWC content}
\usage{
sptf_hwc12(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
This function calculates the hot water extractable carbon content for agricultural soils in the Netherlands. Data from various sources.
}
\references{
Fujita & Ros (2023) Evaluation of pedotransfer functions for soil indicators of BLN. NMI report 1819.N.20. Unpublished data of praktijknetwerk.
}
