% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwc.R
\name{sptf_hwc2}
\alias{sptf_hwc2}
\title{Calculate the HWC content}
\usage{
sptf_hwc2(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
This function calculates the hot water extractable carbon content for agricultural soils in the Netherlands.
}
\references{
Ros (2011) Predicting soil Nitrogen Supply. Unpublished data from pot experiment and incubation experiment.
}
