% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwc.R
\name{sptf_hwc3}
\alias{sptf_hwc3}
\title{Calculate the HWC content}
\usage{
sptf_hwc3(A_C_OF, A_PH_CC)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_PH_CC}{(numeric) The acidity of the soil, pH in CaCl2 (-)}
}
\description{
This function calculates the hot water extractable carbon content for agricultural soils in the Netherlands.
}
\references{
Ros et al. (2011) N-advies op nieuwe leest. NMI-report 1248.N.07, 65 pp.
}
