% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwc.R
\name{sptf_hwc4}
\alias{sptf_hwc4}
\title{Calculate the HWC content}
\usage{
sptf_hwc4(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}
}
\description{
This function calculates the hot water extractable carbon content for grassland soils in the Netherlands.
}
\references{
Van Eekeren & Bokhorst (2009) Beoordeling bodemkwaliteit zandgrond. Een inventarisatie van bodemindicatoren voor de veehouderij. ZvZ-rapport 7, 61 pp.
}
