% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwc.R
\name{sptf_hwc6}
\alias{sptf_hwc6}
\title{Calculate the HWC content}
\usage{
sptf_hwc6(A_C_OF, A_P_AL, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_P_AL}{(numeric) Total phosphate content (mg P2O5/100 g)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
This function calculates the hot water extractable carbon content for agricultural soils in the Netherlands. Data from various years.
}
\references{
Van Balen et al. (2015) Effecten bodem- en structuurverbeteraars. Onderzoek op klei- en zandgrond 2010-2015. PPO-report 693.
}
