% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwc.R
\name{sptf_hwc8}
\alias{sptf_hwc8}
\title{Calculate the HWC content}
\usage{
sptf_hwc8(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}
}
\description{
This function calculates the hot water extractable carbon content for agricultural soils in the Netherlands.
}
\references{
Hanegraaf et al. (2009). Verkenning van bodemsensoren voor de landbouw. Interreg project Bodembreed, NMI project 1382, 93 pp.
}
