% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwc.R
\name{sptf_hwc9}
\alias{sptf_hwc9}
\title{Calculate the HWC content}
\usage{
sptf_hwc9(A_C_OF, A_PH_CC)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_PH_CC}{(numeric) The acidity of the soil, pH in CaCl2 (-)}
}
\description{
This function calculates the hot water extractable carbon content for agricultural soils in the Netherlands. Data from various years.
}
\references{
Hanegraaf et al. (2012) Rekenen met de kwaliteit van organische stof: kengetallen en rekenregels. NMI project 1314.N.09,
}
