% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbial_biomass.R
\name{sptf_mb1}
\alias{sptf_mb1}
\title{Calculate the MB in arable land given the pedo transfer function of Oberholzer and Scheid, 2007}
\usage{
sptf_mb1(A_SOM_LOI, A_CLAY_MI, A_SAND_MI, A_PH_CC)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_PH_CC}{(numeric) Acidity in CaCl2 (-)}
}
\description{
Calculate the MB in arable land given the pedo transfer function of Oberholzer and Scheid, 2007
}
\references{
not available; cited by Greiner, L., Nussbaum, M., Papritz, A., Fraefel, M., Zimmermann, S., Schwab, P., Grêt‐Regamey, A., & Keller, A. (2018). Assessment of soil multi-functionality to support the sustainable use of soil resources on the Swiss Plateau. Geoderma Regional, 14, e00181. https://doi.org/10.1016/j.geodrs.2018.e00181
}
