% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbial_biomass.R
\name{sptf_mb3}
\alias{sptf_mb3}
\title{Calculate the MB in agricultural land treated by tillage and additional SOM given the pedo transfer function of Rui et al. (2016) at experimental site in Australia}
\usage{
sptf_mb3(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the MB in agricultural land treated by tillage and additional SOM given the pedo transfer function of Rui et al. (2016) at experimental site in Australia
}
\references{
Rui, Y., Murphy, D. V., Wang, X., & Hoyle, F. C. (2016). Microbial respiration, but not biomass, responded linearly to increasing light fraction organic matter input: Consequences for carbon sequestration. Scientific Reports, 6(1). https://doi.org/10.1038/srep35496
}
