% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbial_biomass.R
\name{sptf_mb4}
\alias{sptf_mb4}
\title{Calculate the MB in Mulun National Natural Reserve in southwest China}
\usage{
sptf_mb4(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the MB in Mulun National Natural Reserve in southwest China
}
\references{
Qian, Z., Li, Y., Du, H., Wang, K., & Li, D. (2023). Increasing plant species diversity enhances microbial necromass carbon content but does not alter its contribution to soil organic carbon pool in a subtropical forest. Soil Biology & Biochemistry, 187, 109183. https://doi.org/10.1016/j.soilbio.2023.109183
}
