% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbial_biomass.R
\name{sptf_mb5}
\alias{sptf_mb5}
\title{Calculate the MB of agricultural land with different management practices in Tibetan Plateau, China}
\usage{
sptf_mb5(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the MB of agricultural land with different management practices in Tibetan Plateau, China
}
\references{
Li, J. H., Hou, Y. L., Zhang, S. X., Li, W. J., Xu, D. H., Knops, J. M. H., & Shi, X. (2018). Fertilization with nitrogen and/or phosphorus lowers soil organic carbon sequestration in alpine meadows. Land Degradation & Development, 29(6), 1634–1641. https://doi.org/10.1002/ldr.2961
}
