% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwd.R
\name{sptf_mwd1}
\alias{sptf_mwd1}
\title{Calculate the Mean Weight Diameter}
\usage{
sptf_mwd1(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
Calculate the Mean Weight Diameter for agricultural British soils (0-20cm) given the pedotransferfunction of Chaney and Swift (1984)
}
\references{
Chaney and Swift (1984). The influence of organic matter on aggregate stability in some British soils
}
