% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterstableaggregates.R
\name{sptf_mwd10}
\alias{sptf_mwd10}
\title{Calculate the Mean Weight Diamater}
\usage{
sptf_mwd10(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}
}
\description{
Calculate the Mean Weight Diamater for brown calcareous soils (0-5 cm) within the Mediterranean basin in France given the pedotransferfunction of le Bissonnais et al. (2007)
}
\references{
le Bissonnais et al. (2007) Erodibility of Mediterranean vineyard soils: relevant aggregate stability methods and significant soil variables
}
