% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwd.R
\name{sptf_mwd12}
\alias{sptf_mwd12}
\title{Predicting the the Mean Weight Diameter}
\usage{
sptf_mwd12(A_SOM_LOI, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}
}
\description{
Calculate the Mean Weight Diameter (mm) for agricultural topsoils (0-30cm) in India given the pedotransferfunction of Bhattacharya et al. (2021)
}
\references{
Bhattacharya et al. (2021) Prediction of mean weight diameter of soil using machine learning approaches
}
