% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterstableaggregates.R
\name{sptf_mwd13}
\alias{sptf_mwd13}
\title{Predicting the the Mean Weight Diameter for soils in Morroco}
\usage{
sptf_mwd13(A_CLAY_MI, A_SOM_LOI)
}
\arguments{
\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SOM_LOI}{(numeric) The soil organic matter content (\%).}
}
\description{
Calculate the Mean Weight Diameter (mm) for agricultural soils (0-40 cm) in Morroco given the pedotransferfunction of Bouslihim et al. (2021)
}
\references{
Bouslihim et al. (2021) Machine learning approaches for the prediction of soil aggregate stability
}
