% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwd.R
\name{sptf_mwd15}
\alias{sptf_mwd15}
\title{Predicting the the Mean Weight Diameter for soils}
\usage{
sptf_mwd15(A_C_OF, A_PH_WA)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_PH_WA}{(numeric) Acidity in water (-)}
}
\description{
Calculate the Mean Weight Diameter for agricultural topsoils (0-10cm) in Belgium given the pedotransferfunction of Shi et al. (2020)
}
\references{
Shi et al. (2020). Vis-NIR spectroscopic assessment of soil aggregate stability and aggregate size distribution in the Belgian Loam Belt
}
