% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwd.R
\name{sptf_mwd16}
\alias{sptf_mwd16}
\title{Predicting the the Mean Weight Diameter for soils}
\usage{
sptf_mwd16(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}
}
\description{
Calculate the Mean Weight Diamater for agricultural topsoils (0-28cm) in France.
}
\references{
Chenu et al. (2000) Organic Matter Influence on Clay Wettability and Soil Aggregate Stability
}
