% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwd.R
\name{sptf_mwd17}
\alias{sptf_mwd17}
\title{Calculate the Mean Weight Diameter}
\usage{
sptf_mwd17(A_C_OF, A_CLAY_MI, A_SILT_MI, A_CACO3_IF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_CACO3_IF}{(numeric) Carbonated Lime inorganic soil fraction (\%)}
}
\description{
Calculate the Mean Weight Diameter for agricultural soils (0-10cm) in Tunesia given the pedotransferfunction of Gomez et al. (2013)
}
\references{
Gomez et al., (2013) Laboratory Vis–NIR spectroscopy as an alternative method for estimating the soil aggregate stability indexes of Mediterranean soils
}
