% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwd.R
\name{sptf_mwd18}
\alias{sptf_mwd18}
\title{Calculate the Mean Weight Diameter}
\usage{
sptf_mwd18(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the Mean Weight Diameter for Qiyang, Hunan province in China
}
\references{
Mustafa, A., Xu, M., Shah, S. a. A., Abrar, M., Sun, N., Wang, B., Cai, Z., Saeed, Q., Naveed, M., Mehmood, K., & Núñez‐Delgado, A. (2020c). Soil aggregation and soil aggregate stability regulate organic carbon and nitrogen storage in a red soil of southern China. Journal of Environmental Management, 270, 110894. https://doi.org/10.1016/j.jenvman.2020.110894
}
