% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwd.R
\name{sptf_mwd19}
\alias{sptf_mwd19}
\title{Calculate the Mean Weight Diameter}
\usage{
sptf_mwd19(A_SOM_LOI, A_CLAY_MI, A_C_OF)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the Mean Weight Diameter for agricultural soils in Fars province, which is located in the south-central region of Iran
}
\references{
Mina, M., Rezaei, M., Sameni, A., Riksen, M., & Ritsema, C. (2023). Estimating the indices of soil erodibility to wind erosion using pedo- and spectro-transfer functions in calcareous soils. Geoderma, 438, 116612. https://doi.org/10.1016/j.geoderma.2023.116612
}
