% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwd.R
\name{sptf_mwd2}
\alias{sptf_mwd2}
\title{Calculate the Mean Weight Diamater}
\usage{
sptf_mwd2(A_C_OF, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}
}
\description{
Calculate the Mean Weight Diamater for agricultural top soils (0-20cm) in New Zealand given the pedotransferfunction of Haynes et al. (1991)
}
\references{
Haynes et al. (1991) Influence of mixed cropping rotations (pasture-arable) on organic matter content, water-stable aggregation and clod porosity in a group of soils
}
