% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwd.R
\name{sptf_mwd3}
\alias{sptf_mwd3}
\title{Calculate the Mean Weight Diameter}
\usage{
sptf_mwd3(A_C_OF, A_CEC_CO, A_CLAY_MI, A_SILT_MI, A_PH_WA, A_CACO3_IF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CEC_CO}{(numeric) Cation Exchange Capacity (mmol+/kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_PH_WA}{(numeric) Acidity in water (-)}

\item{A_CACO3_IF}{(numeric) Carbonated Lime inorganic soil fraction (\%)}
}
\description{
Calculate the Mean Weight Diameter for soils (0-20cm) in Northwestern Algeria given the pedotransferfunction of Saidi et al. (2015)
}
\references{
Saidi et al. (2015). Using Pedotransfer Functions to Assess Aggregate Stability: Application to the Lower Cheliff Soils, Algeria
}
