% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwd.R
\name{sptf_mwd6}
\alias{sptf_mwd6}
\title{Calculate the Mean Weight Diamater}
\usage{
sptf_mwd6(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}
}
\description{
Calculate the Mean Weight Diamater for agricultural soils (0-20cm) in the Loess Plateau of China given the pedotransferfunction of Gao et al. (2019)
}
\references{
Gao et al. (2019). Soil wet aggregate distribution and pore size distribution under different tillage systems after 16 years in the Loess Plateau of China
}
