% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwd.R
\name{sptf_mwd8}
\alias{sptf_mwd8}
\title{Calculate the Mean Weight Diameter}
\usage{
sptf_mwd8(A_SOM_LOI, A_CLAY_MI, A_SILT_MI, A_PH_WA, A_CACO3_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_PH_WA}{(numeric) Acidity in water (-)}

\item{A_CACO3_MI}{(numeric) The calcium carbonate content of the soil (\%)}
}
\description{
Calculate the Mean Weight Diameter for agricultural soils (0-20cm) for Algeria, extended withh data from published results from other countries, given the pedotransferfunction of Hamel et al. (2021)
}
\references{
Hamel et al. (2021). Evaluation of soil aggregate stability in Algerian northwestern soils using pedotransfer functions and artificial neural networks
}
