% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterstableaggregates.R
\name{sptf_mwd9}
\alias{sptf_mwd9}
\title{Calculate the Mean Weight Diamater}
\usage{
sptf_mwd9(A_C_OF, A_CLAY_MI, A_SILT_MI, A_PH_WA, B_LU_PTFCLASS)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}

\item{A_PH_WA}{(numeric) The acidity of the soil, pH in water (-)}

\item{B_LU_PTFCLASS}{(character) The land use category (options: agriculture, grassland, cropland, forest, nature)}
}
\description{
Calculate the Mean Weight Diamater for variable land uses in France (0-20cm) given the pedotransferfunction of Clergue et al. (2023)
}
\references{
Clergue et al. (2023). Estimating soil aggregate stability with infrared spectroscopy and pedotransfer functions
}
