% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw10}
\alias{sptf_paw10}
\title{Calculate the Plant Available Water (PAW) given the pedotransferfunction of Rawls et al.1982}
\usage{
sptf_paw10(A_C_OF, A_CLAY_MI, A_SILT_MI, A_SAND_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_SAND_MI}{(numeric) Sand content (\%)}
}
\description{
Calculate the Plant Available Water (PAW) given the pedotransferfunction of Rawls et al.1982
}
\details{
mp_fc must be either 10 or 33
}
\references{
Rawls et al.(1982) Estimation of soil water properties. Trans. ASAE 25, 1316–1320.
}
