% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw11}
\alias{sptf_paw11}
\title{Calculate the Plant Available Water (PAW) given the pedotransferfunction of Campbell & Shiozawa 1992
Note: Calculation of whc parameters may be wrong. The calculated water content is out of normal range. Original literature can not be found on web.}
\usage{
sptf_paw11(A_C_OF, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}
}
\description{
Calculate the Plant Available Water (PAW) given the pedotransferfunction of Campbell & Shiozawa 1992
Note: Calculation of whc parameters may be wrong. The calculated water content is out of normal range. Original literature can not be found on web.
}
\details{
mp_fc must be either 10 or 33
}
\references{
Campbell & Shiozawa (1992) Prediction of hydraulic properties of soils using particle-size distribution and bulk density data
}
