% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw13}
\alias{sptf_paw13}
\title{Calculate the Plant Available Water (PAW) given the pedotransferfunction of Tian et al. 2021}
\usage{
sptf_paw13(A_C_OF, A_CLAY_MI, A_SAND_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SAND_MI}{(numeric) Sand content (\%)}
}
\description{
Calculate the Plant Available Water (PAW) given the pedotransferfunction of Tian et al. 2021
}
\references{
Tian et al. (2021) New pedotransfer functions for soil water retention curves that better account for bulk density effects
}
