% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw15}
\alias{sptf_paw15}
\title{Calculate the Plant Available Water (PAW) given the pedotransferfunction of Wösten et al.2001}
\usage{
sptf_paw15(A_SOM_LOI, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}
}
\description{
Calculate the Plant Available Water (PAW) given the pedotransferfunction of Wösten et al.2001
}
\references{
Wösten, J. H. M., Veerman, G. ., de Groot, W. J., & Stolte, J. (2001). Waterretentie en doorlatendheidskarakteristieken van boven- en ondergronden in Nederland: de Staringreeks. Alterra Rapport, 153, 86. https://edepot.wur.nl/43272
}
