% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw17}
\alias{sptf_paw17}
\title{Calculate the Plant Available Water (PAW) given the pedotransferfunction of Lal et al.1978 (Table V & VI), for group I}
\usage{
sptf_paw17(A_CLAY_MI, A_C_OF)
}
\arguments{
\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the Plant Available Water (PAW) given the pedotransferfunction of Lal et al.1978 (Table V & VI), for group I
}
\references{
Lal, R. (1978). Physical properties and moisture retention characteristics of some nigerian soils. Geoderma, 21(3), 209–223. https://doi.org/10.1016/0016-7061(78)90028-9
}
