% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw18}
\alias{sptf_paw18}
\title{Calculate the Plant Available Water (PAW) given the pedotransfer function of chen et al., 2005 about soil in SHanxi and Henan province in China}
\usage{
sptf_paw18(A_C_OF, A_SAND_MI, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}
}
\description{
Calculate the Plant Available Water (PAW) given the pedotransfer function of chen et al., 2005 about soil in SHanxi and Henan province in China
}
\references{
王志强.(2003).科尔沁沙地土壤水力特性的推算(硕士学位论文,内蒙古农业大学).https://kns.cnki.net/KCMS/detail/detail.aspx?dbname=CMFD9904&filename=2003085422.nh
}
