% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw19}
\alias{sptf_paw19}
\title{Calculate the Plant Available Water (PAW) given the pedotransferfunction of Dijkerman for soil in Sierra Leone}
\usage{
sptf_paw19(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Calculate the Plant Available Water (PAW) given the pedotransferfunction of Dijkerman for soil in Sierra Leone
}
\references{
Dijkerman, J. (1988). An Ustult-Aquult-Tropept catena in Sierra Leone, West Africa, II. Land qualities and land evaluation. Geoderma, 42(1), 29–49. https://doi.org/10.1016/0016-7061(88)90021-3
}
