% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw20}
\alias{sptf_paw20}
\title{Calculate the Plant Available Water (PAW) given the pedotransferfunction of Aina and Periaswamy (1985)  for soil in Nigeria}
\usage{
sptf_paw20(A_C_OF, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}
}
\description{
Calculate the Plant Available Water (PAW) given the pedotransferfunction of Aina and Periaswamy (1985)  for soil in Nigeria
}
\references{
Aina, P. O., & Periaswamy, S. P. (1985). ESTIMATING AVAILABLE WATER-HOLDING CAPACITY OF WESTERN NIGERIAN SOILS FROM SOIL TEXTURE AND BULK DENSITY, USING CORE AND SIEVED SAMPLES. Soil Science, 140(1), 55–58. https://doi.org/10.1097/00010694-198507000-00007
}
