% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw21}
\alias{sptf_paw21}
\title{Calculate the Plant Available Water (PAW) given the pedotransferfunction of Saxton & Rawl et al.2006}
\usage{
sptf_paw21(A_SOM_LOI, A_SAND_MI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Calculate the Plant Available Water (PAW) given the pedotransferfunction of Saxton & Rawl et al.2006
}
\references{
Saxton, K. E., & Rawls, W. J. (2006). Soil water characteristic estimates by texture and organic matter for hydrologic solutions. Soil Science Society of America Journal, 70(5), 1569–1578. https://doi.org/10.2136/sssaj2005.0117
}
