% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw22}
\alias{sptf_paw22}
\title{Calculate the Plant Available Water (PAW) given the pedotransferfunction of da Costa et al.2013 for surface soil}
\usage{
sptf_paw22(A_SOM_LOI, A_SAND_MI, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}
}
\description{
Calculate the Plant Available Water (PAW) given the pedotransferfunction of da Costa et al.2013 for surface soil
}
\references{
Da Costa, A., Albuquerque, J. A., De Almeida, J. A., Da Costa, A., & Luciano, R. V. (2013). Pedotransfer functions to estimate retention and availability of water in soils of the state of Santa Catarina, Brazil. Revista Brasileira De Ciencia Do Solo, 37(4), 889–910. https://doi.org/10.1590/s0100-06832013000400007
}
