% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw24}
\alias{sptf_paw24}
\title{Calculate the Plant Available Water (PAW) given the pedotransferfunction of Rawl et al.2003 for soil data from US national soil characterization database}
\usage{
sptf_paw24(A_C_OF, A_SAND_MI, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Calculate the Plant Available Water (PAW) given the pedotransferfunction of Rawl et al.2003 for soil data from US national soil characterization database
}
\references{
Rawls, W. J., Pachepsky, Y., Ritchie, J. C., Sobecki, T. M., & Bloodworth, H. (2003b). Effect of soil organic carbon on soil water retention. Geoderma, 116(1–2), 61–76. https://doi.org/10.1016/s0016-7061(03)00094-6
}
