% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw26}
\alias{sptf_paw26}
\title{Calculate the Plant Available Water (PAW) given the pedotransfer function from model 6 of Katterer et al.2007 for soil data from Swedish subsoil}
\usage{
sptf_paw26(A_C_OF, A_SAND_MI, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Calculate the Plant Available Water (PAW) given the pedotransfer function from model 6 of Katterer et al.2007 for soil data from Swedish subsoil
}
\references{
Kätterer, T., Andrén, O., & Jansson, P. (2006). Pedotransfer functions for estimating plant available water and bulk density in Swedish agricultural soils. Acta Agriculturae Scandinavica Section B-soil and Plant Science, 56(4), 263–276. https://doi.org/10.1080/09064710500310170
}
