% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw27}
\alias{sptf_paw27}
\title{Calculate the Plant Available Water (PAW) given the pedotransfer function of Dobarco et al.2019 for soil data from France}
\usage{
sptf_paw27(A_C_OF, A_SAND_MI, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Calculate the Plant Available Water (PAW) given the pedotransfer function of Dobarco et al.2019 for soil data from France
}
\references{
Dobarco, M. Ŕ., Cousin, I., Bas, C. L., & Martín, M. (2019). Pedotransfer functions for predicting available water capacity in French soils, their applicability domain and associated uncertainty. Geoderma, 336, 81–95. https://doi.org/10.1016/j.geoderma.2018.08.022
}
