% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw28}
\alias{sptf_paw28}
\title{Calculate the Plant Available Water (PAW) given the pedotransfer function of Gupta and Larson (1979) for soil data from United States}
\usage{
sptf_paw28(A_C_OF, A_SOM_LOI, A_SAND_MI, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}
}
\description{
Calculate the Plant Available Water (PAW) given the pedotransfer function of Gupta and Larson (1979) for soil data from United States
}
\references{
Gupta, S. C., & Larson, W. E. (1979). Estimating soil water retention characteristics from particle size distribution, organic matter percent, and bulk density. Water Resources Research, 15(6), 1633–1635. https://doi.org/10.1029/wr015i006p01633
}
