% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw29}
\alias{sptf_paw29}
\title{Calculate the Plant Available Water (PAW) given the pedotransfer function of Yost and Hartemink (2019) for soil data from United States}
\usage{
sptf_paw29(A_C_OF, A_SAND_MI, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}
}
\description{
Calculate the Plant Available Water (PAW) given the pedotransfer function of Yost and Hartemink (2019) for soil data from United States
}
\references{
Yost, J. L., & Hartemink, A. E. (2019). Effects of carbon on moisture storage in soils of the Wisconsin Central Sands, USA. European Journal of Soil Science, 70(3), 565–577. https://doi.org/10.1111/ejss.12776
}
