% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw3}
\alias{sptf_paw3}
\title{Calculate the Plant Available Water (PAW) given the pedotransferfunction of Saxton et al.1986}
\usage{
sptf_paw3(A_SAND_MI, A_CLAY_MI)
}
\arguments{
\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Calculate the Plant Available Water (PAW) given the pedotransferfunction of Saxton et al.1986
}
\details{
mp_fc must be larger than 10
}
\references{
Saxton et al.(1986) Estimating Generalized Soil-water Characteristics from Texture
}
