% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw30}
\alias{sptf_paw30}
\title{Calculate the Plant Available Water (PAW) given the pedotransfer function of chen et al., 2005 about soil in SHanxi and Henan province in China}
\usage{
sptf_paw30(A_SAND_MI, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}
}
\description{
Calculate the Plant Available Water (PAW) given the pedotransfer function of chen et al., 2005 about soil in SHanxi and Henan province in China
}
\references{
陈晓燕,陆桂华,秦福兴,等. 土壤传递函数法在确定田间持水量中的应用[J]. 河海大学学报（自然科学版）,2005,33(2):170-172. DOI:10.3321/j.issn:1000-1980.2005.02.013.
}
