% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw31}
\alias{sptf_paw31}
\title{Calculate the Plant Available Water (PAW) given the pedotransfer function about soil in the Balaguer River basin in China}
\usage{
sptf_paw31(A_C_OF, A_SOM_LOI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}
}
\description{
Calculate the Plant Available Water (PAW) given the pedotransfer function about soil in the Balaguer River basin in China
}
\references{
房丽晶,高瑞忠,刘廷玺,张阿龙 & 王喜喜.(2020).巴拉格尔河流域土壤传递函数构建与评估. 干旱区研究(05),1156-1165. doi:10.13866/j.azr.2020.05.08.
}
