% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw33}
\alias{sptf_paw33}
\title{Calculate the Plant Available Water (PAW) given the pedotransfer function of chen et al., 2005 about soil in Upper Yangtze River basin in China}
\usage{
sptf_paw33(A_C_OF, A_SOM_LOI, A_SAND_MI, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Calculate the Plant Available Water (PAW) given the pedotransfer function of chen et al., 2005 about soil in Upper Yangtze River basin in China
}
\references{
邓建波.(2020).不同土地利用类型土壤水力学参数的传递函数研究(硕士学位论文,四川农业大学).https://kns.cnki.net/KCMS/detail/detail.aspx?dbname=CMFD202201&filename=1021681965.nh
}
