% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw34}
\alias{sptf_paw34}
\title{Calculate the Plant Available Water (PAW) given the pedotransfer function of chen et al., 2005 about soil in SHanxi and Henan province in China}
\usage{
sptf_paw34(A_C_OF, A_SAND_MI, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Calculate the Plant Available Water (PAW) given the pedotransfer function of chen et al., 2005 about soil in SHanxi and Henan province in China
}
\references{
朱安宁,张佳宝,程竹华.(2003).轻质土壤水分特征曲线估计的简便方法. 土壤通报(04),253-258. doi:10.19336/j.cnki.trtb.2003.04.004.
}
