% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw35}
\alias{sptf_paw35}
\title{Estimate PAW based on water retention curve parameters by Sun et al. (2022)}
\usage{
sptf_paw35(A_C_OF, A_SOM_LOI, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}
}
\description{
Estimate PAW based on water retention curve parameters by Sun et al. (2022)
}
\references{
孙志祥,邓建波,吕玉娟等.长江上游低山丘陵区土壤水分特征曲线传递函数研究[J].灌溉排水学报,2022,41(06):97-104.DOI:10.13522/j.cnki.ggps.2021609.
}
