% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw36}
\alias{sptf_paw36}
\title{Estimate PAW based on water retention curve parameters by Cui et al. (2021)}
\usage{
sptf_paw36(A_C_OF, A_SOM_LOI, A_SAND_MI, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Estimate PAW based on water retention curve parameters by Cui et al. (2021)
}
\references{
崔俊芳,邓建波,刘传栋 & 唐翔宇.(2021).若尔盖高寒草甸表层土壤水分特征曲线传递函数研究. 山地学报(04),483-494. doi:10.16089/j.cnki.1008-2786.000613.
}
