% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw5}
\alias{sptf_paw5}
\title{Calculate the Plant Available Water (PAW) given the pedotransferfunction of Wösten et al. 1999}
\usage{
sptf_paw5(A_SOM_LOI, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}
}
\description{
Calculate the Plant Available Water (PAW) given the pedotransferfunction of Wösten et al. 1999
}
\references{
Wösten et al. (1999) Development and use of a database of hydraulic properties of European soils
}
