% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw.R
\name{sptf_paw8}
\alias{sptf_paw8}
\title{Calculate the Plant Available Water (PAW) given the pedotransferfunction of Weynants et al. 2009}
\usage{
sptf_paw8(A_C_OF, A_CLAY_MI, A_SAND_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SAND_MI}{(numeric) Sand content (\%)}
}
\description{
Calculate the Plant Available Water (PAW) given the pedotransferfunction of Weynants et al. 2009
}
\references{
Weynants et al.(2009) Revisiting Vereecken Pedotransfer Functions: Introducing a Closed-Form Hydraulic Model
}
