% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{sptf_ph1}
\alias{sptf_ph1}
\title{Calculate the pH-water value from pH-KCL}
\usage{
sptf_ph1(A_PH_KCL)
}
\arguments{
\item{A_PH_KCL}{(numeric) The acidity of the soil, pH in KCL (-)}
}
\description{
This function calculates the pH extracted with water from the pH-KCL.
}
\references{
Wosten et al. (1997) Bodemkundige vertaalfuncties bij SC-DLO; state of the art. https://edepot.wur.nl/299664
}
