% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phbc.R
\name{sptf_phbc1}
\alias{sptf_phbc1}
\title{Calculate the soil pH buffering capacity}
\usage{
sptf_phbc1(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
This function calculates the capacity of soils to buffer pH changes.
}
\details{
pHBC determined by titration according to method of Aitken & Moody (1994)
}
\references{
McBratney et al. (2002). From pedotransfer functions to soil inference systems.
}
