% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phbc.R
\name{sptf_phbc2}
\alias{sptf_phbc2}
\title{Calculate the soil pH buffering capacity}
\usage{
sptf_phbc2(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}
}
\description{
This function calculates the capacity of soils (0-10cm) in Australia to buffer pH changes in response to changes in SOC.
}
\details{
pHBC determined by titration according to method of Aitken & Moody (1994)
}
\references{
Wong et al. (2013) Development of buffer methods and evaluation of pedotransfer functions to estimate pH buffer capacity of highly weathered soils
}
