% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phbc.R
\name{sptf_phbc3}
\alias{sptf_phbc3}
\title{Calculate the soil pH buffering capacity}
\usage{
sptf_phbc3(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
This function calculates the capacity of soils to buffer pH changes in response to changes in SOC.
}
\details{
pHBC determined by titration according to method of Aitken & Moody (1994). Datasets derived from literature.
}
\references{
Wong et al. (2013) Development of buffer methods and evaluation of pedotransfer functions to estimate pH buffer capacity of highly weathered soils
}
