% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilproperties.R
\name{sptf_phbc5}
\alias{sptf_phbc5}
\title{Calculate the soil pH buffering capacity}
\usage{
sptf_phbc5(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}
}
\description{
This function calculates the capacity of soils in Georgia to buffer pH changes in response to changes in SOC.
}
\details{
pHBC determined by titration according to method of Aitken & Moody (1994) using CaOH2.
}
\references{
Weaver et al. (2004) Mapping Soil pH Buffering Capacity of Selected Fields in the Coastal Plain
}
