% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phbc.R
\name{sptf_phbc6}
\alias{sptf_phbc6}
\title{Calculate the soil pH buffering capacity}
\usage{
sptf_phbc6(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
This function calculates the capacity of soils to buffer pH changes in agricultral topsoil in Queensland (Australia) in response to changes in SOC.
}
\details{
pHBC determined by titration according to method of Aitken & Moody (1994) using CaOH2.
}
\references{
Aitken (1990) Lime Requirement of Acidic Queensland Soils. I. Relationships between Soil Properties and pH Buffer Capacity
}
