% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilproperties.R
\name{sptf_phbc7}
\alias{sptf_phbc7}
\title{Calculate the soil pH buffering capacity}
\usage{
sptf_phbc7(A_C_OF, A_CLAY_MI, A_PH_WA)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_PH_WA}{(numeric) The acidity of the soil, pH in water (-)}
}
\description{
This function calculates the capacity of soils to buffer pH changes in response to changes in SOC, clay and pH.
}
\details{
pHBC determined by titration according to using CaOH2 titration method.
}
\references{
Owusu et al. (1995) Comparative study of selected lime requirement methods for some acid Ghanaian soils
}
