% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phbc.R
\name{sptf_phbc8}
\alias{sptf_phbc8}
\title{Calculate the soil pH buffering capacity}
\usage{
sptf_phbc8(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
This function calculates the capacity of soils to buffer pH changes in response to changes in SOC and clay.
}
\details{
pHBC determined by titration according to using CaOH2 titration method.
}
\references{
Joret et al. (1934) Líappreciation des besoins en chaux des sols de limon díaprËs leur Ètat de saturation en bases Èchangeables.
}
