% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmn.R
\name{sptf_pmn1}
\alias{sptf_pmn1}
\title{Calculate the PMN given the pedo transfer function of Narteh and Sahrawat (1997)}
\usage{
sptf_pmn1(A_C_OF, A_CLAY_MI, A_N_RT, A_PH_CC)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_N_RT}{(numeric) The organic nitrogen content of the soil in mg N / kg}

\item{A_PH_CC}{(numeric) The acidity of the soil, measured in 0.01M CaCl2 (-)}
}
\description{
Calculate the PMN given the pedo transfer function of Narteh and Sahrawat (1997)
}
\references{
Narteh and Sahrawat (1997) Potentially Mineralizable Nitrogen in West African Lowland Rice Soils. Geoderma, 76, 145-154
}
