% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmn.R
\name{sptf_pmn10}
\alias{sptf_pmn10}
\title{Calculate the PMN given the pedo transfer function of Rasiah (1995), based on dataset of Simard & N'dayegamiye (1993)}
\usage{
sptf_pmn10(A_N_RT, t = 7)
}
\arguments{
\item{A_N_RT}{(numeric) Total nitrogen content (mg N/ kg)}

\item{t}{(numeric) Length of incubation period (days).}
}
\description{
Calculate the PMN given the pedo transfer function of Rasiah (1995), based on dataset of Simard & N'dayegamiye (1993)
}
\references{
Cimard & N'dayegamiye (1993) Nitrogen -ninerahzation potential of meadow soils. Can. J. Soil Sci. 73:27-38
}
