% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmn.R
\name{sptf_pmn12}
\alias{sptf_pmn12}
\title{Calculate the Potentially Mineralizable N (PMN) in mg/kg}
\usage{
sptf_pmn12(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
This function calculates the PMN content for agricultural soils in the Netherlands. PMN refers to anearobic mineralizable N pool, 40 degrees for 7 days.
}
\references{
Ros (2011) Predicting soil Nitrogen Supply. Unpublished data from pot experiment and incubation experiment.
}
