% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmn.R
\name{sptf_pmn15}
\alias{sptf_pmn15}
\title{Calculate the Potentially Mineralizable N (PMN) in mg/kg}
\usage{
sptf_pmn15(A_C_OF, A_P_AL, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_P_AL}{(numeric) Total phosphate content (mg P2O5/100 g)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
This function calculates the PMN content for agricultural soils in the Netherlands. PMN refers to anearobic mineralizable N pool, 40 degrees for 7 days.
Data from various years.
}
\references{
Van Balen et al. (2015) Effecten bodem- en structuurverbeteraars. Onderzoek op klei- en zandgrond 2010-2015. PPO-report 693.
}
