% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmn.R
\name{sptf_pmn18}
\alias{sptf_pmn18}
\title{Calculate the Potentially Mineralizable N (PMN) in mg/kg}
\usage{
sptf_pmn18(A_C_OF, A_PH_CC)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_PH_CC}{(numeric) Acidity in CaCl2 (-)}
}
\description{
This function calculates the PMN content for agricultural soils in the Netherlands. PMN refers to anearobic mineralizable N pool, 40 degrees for 7 days.
}
\references{
Hanegraaf et al. (2009). De afbraaksnelheid van organische stof in Drenthe. NMI project 972.N.09, 93 pp.
}
