% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmn.R
\name{sptf_pmn19}
\alias{sptf_pmn19}
\title{Calculate the HWC content}
\usage{
sptf_pmn19(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
This function calculates the hot water extractable carbon content for agricultural soils in the Netherlands. Data from various years.
}
\references{
Van Eekeren & Bokhorst (2009) Beoordeling bodemkwaliteit zandgrond. Een inventarisatie van bodemindicatoren voor de veehouderij. ZvZ-rapport 7, 61 pp. Unpublished data.
}
