% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmn.R
\name{sptf_pmn2}
\alias{sptf_pmn2}
\title{Calculate the PMN given the pedo transfer function of Zou et al (2018)}
\usage{
sptf_pmn2(A_C_OF, A_SILT_MI, A_N_RT, A_PH_CC, t = 7)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}

\item{A_N_RT}{(numeric) The organic nitrogen content of the soil in mg N / kg}

\item{A_PH_CC}{(numeric) The acidity of the soil, measured in 0.01M CaCl2 (-)}

\item{t}{(numeric) Length of incubation period (days). The empirical data is ~35 days, thus it should not exceed this value too much.}
}
\description{
Calculate the PMN given the pedo transfer function of Zou et al (2018)
}
\references{
Zou et al. (2018) Pedo-Transfer Functions to Estimate Kinetic Parameters for Anaerobic Soil Nitrogen Mineralization. Open Journal of Soil Science, 8, 75-86. doi: 10.4236/ojss.2018.82006.
}
