% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmn.R
\name{sptf_pmn21}
\alias{sptf_pmn21}
\title{Calculate the PMN given the pedo transfer function of Zhang et al. (2017), based on incubation data in China}
\usage{
sptf_pmn21(A_C_OF, t = 7)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{t}{(numeric) Length of incubation period (days).}
}
\description{
Calculate the PMN given the pedo transfer function of Zhang et al. (2017), based on incubation data in China
}
\references{
Zhang, Y., Wang, X., Duan, P., Cong, Y., An, T., Yu, N., Zou, H., Dang, X., An, J., Fan, Q., & Zhang, Y. (2017). Evaluation and simulation of nitrogen mineralization of paddy soils in Mollisols area of Northeast China under waterlogged incubation. PLOS ONE, 12(2), e0171022. https://doi.org/10.1371/journal.pone.0171022
}
