% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmn.R
\name{sptf_pmn22}
\alias{sptf_pmn22}
\title{Calculate the PMN given the pedo transfer function of Ve et al. (2004), based on data in Phlippine land growing rice
The mineralized N is measured by the accumulation of NH4-N}
\usage{
sptf_pmn22(A_C_OF, A_N_RT, t = 7)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_N_RT}{(numeric) Total nitrogen content (mg N/ kg)}

\item{t}{(numeric) Length of incubation period (days).}
}
\description{
Calculate the PMN given the pedo transfer function of Ve et al. (2004), based on data in Phlippine land growing rice
The mineralized N is measured by the accumulation of NH4-N
}
\references{
Ve, N. B., Olk, D. C., & Cassman, K. G. (2004). Characterization of humic acid fractions improves estimates of nitrogen mineralization kinetics for lowland rice soils. Soil Science Society of America Journal, 68(4), 1266–1277. https://doi.org/10.2136/sssaj2004.1266
}
