% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmn.R
\name{sptf_pmn23}
\alias{sptf_pmn23}
\title{Calculate the PMN given the pedo transfer function of Sahrawat (1983), based on incubation data of Phlippine land growing rice
The mineralized N is measured by released ammonium}
\usage{
sptf_pmn23(A_C_OF, A_N_RT, A_CN_FR, A_PH_WA, A_CEC_CO, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_N_RT}{(numeric) Total nitrogen content (mg N/ kg)}

\item{A_CN_FR}{(numeric) Carbon / Nitrogen ratio (ratio)}

\item{A_PH_WA}{(numeric) Acidity in water (-)}

\item{A_CEC_CO}{(numeric) Cation Exchange Capacity (mmol+/kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Calculate the PMN given the pedo transfer function of Sahrawat (1983), based on incubation data of Phlippine land growing rice
The mineralized N is measured by released ammonium
}
\references{
Sahrawat, K. L. (1983). Mineralization of soil organic nitrogen under waterlogged conditions in relation to other properties of tropical rice soils. Soil Research, 21(2), 133. https://doi.org/10.1071/sr9830133
}
