% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmn.R
\name{sptf_pmn24}
\alias{sptf_pmn24}
\title{Calculate the PMN given the pedo transfer function of Kadono et al. (2008), based on data from Ukraine and Kazakhastan}
\usage{
sptf_pmn24(A_N_RT, t = 7)
}
\arguments{
\item{A_N_RT}{(numeric) Total nitrogen content (mg N/ kg)}

\item{t}{(numeric) Length of incubation period (days).}
}
\description{
Calculate the PMN given the pedo transfer function of Kadono et al. (2008), based on data from Ukraine and Kazakhastan
}
\references{
Kadono, A., Funakawa, S., & Kosaki, T. (2008). Factors controlling mineralization of soil organic matter in the Eurasian steppe. Soil Biology & Biochemistry, 40(4), 947–955. https://doi.org/10.1016/j.soilbio.2007.11.015
}
