% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmn.R
\name{sptf_pmn25}
\alias{sptf_pmn25}
\title{Calculate the PMN given the pedo transfer function of Sharifi et al. (2007), based on data from from 17 existing field experiments in New Brunswick, Quebec, Manitoba, and Saskatchewan, Canada, and Maine, USA.}
\usage{
sptf_pmn25(A_C_OF, t = 7)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{t}{(numeric) Length of incubation period (days).}
}
\description{
Calculate the PMN given the pedo transfer function of Sharifi et al. (2007), based on data from from 17 existing field experiments in New Brunswick, Quebec, Manitoba, and Saskatchewan, Canada, and Maine, USA.
}
\references{
Sharifi, M., Zebarth, B. J., Burton, D. L., Grant, C. A., & Cooper, J. (2007). Evaluation of some indices of potentially mineralizable nitrogen in soil. Soil Science Society of America Journal, 71(4), 1233–1239. https://doi.org/10.2136/sssaj2006.0265
}
