% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmn.R
\name{sptf_pmn3}
\alias{sptf_pmn3}
\title{Calculate the PMN given the pedo transfer function of Ros & Fujita 2020 for Dutch soils}
\usage{
sptf_pmn3(A_N_RT, A_CLAY_MI)
}
\arguments{
\item{A_N_RT}{(numeric) The organic nitrogen content of the soil in mg N / kg}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
Calculate the PMN given the pedo transfer function of Ros & Fujita 2020 for Dutch soils
}
\references{
Ros & Fujita (2020) FACTSHEET actools: Potential Mineralizable N Index v.0.4.
}
