% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmn.R
\name{sptf_pmn4}
\alias{sptf_pmn4}
\title{Calculate the PMN given the pedo transfer function of Ros & Fujita 2020 for global soils}
\usage{
sptf_pmn4(A_C_OF, A_N_RT, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_N_RT}{(numeric) Total nitrogen content (mg N/ kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Calculate the PMN given the pedo transfer function of Ros & Fujita 2020 for global soils
}
\references{
Ros & Fujita (2020) FACTSHEET actools: Potential Mineralizable N Index v.0.4.
}
