% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmn.R
\name{sptf_pmn5}
\alias{sptf_pmn5}
\title{Calculate the PMN given the pedo transfer function of Rasiah (1995), based on dataset of Stanford and Smith (1972)}
\usage{
sptf_pmn5(A_N_RT, A_C_OF, A_CEC_CO, t = 7)
}
\arguments{
\item{A_N_RT}{(numeric) Total nitrogen content (mg N/ kg)}

\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CEC_CO}{(numeric) Cation Exchange Capacity (mmol+/kg)}

\item{t}{(numeric) Length of incubation period (days).}
}
\description{
Calculate the PMN given the pedo transfer function of Rasiah (1995), based on dataset of Stanford and Smith (1972)
}
\references{
Rasiah (1995). Comparison of pedotransfer functions to predict nitrogen-mineralization parameters of one- and two-pool models. Communications in Soil Science and Plant Analysis 26, 1873–1884.

Stanford and Smith (1972) Nitrogen mineralization potentials of soils. Soil Sci. Soc. Amer. J. 36:465-472.
}
