% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmn.R
\name{sptf_pmn6}
\alias{sptf_pmn6}
\title{Calculate the PMN given the pedo transfer function of Rasiah (1995), based on Cabrera & Kissel (1988)}
\usage{
sptf_pmn6(A_N_RT, A_C_OF, A_CLAY_MI, A_SILT_MI, A_PH_CC, t = 7, RES = 1)
}
\arguments{
\item{A_N_RT}{(numeric) The organic nitrogen content of the soil in mg N / kg}

\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}

\item{A_PH_CC}{(numeric) The pH-CaCl2 of the soil}

\item{t}{(numeric) Length of incubation period (days).}

\item{RES}{(numeric) whether residue is not removed (1) or removed (0)}
}
\description{
Calculate the PMN given the pedo transfer function of Rasiah (1995), based on Cabrera & Kissel (1988)
}
\references{
Rasiah (1995). Comparison of pedotransfer functions to predict nitrogen-mineralization parameters of one- and two-pool models. Communications in Soil Science and Plant Analysis 26, 1873–1884.
}
