% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmn.R
\name{sptf_pmn7}
\alias{sptf_pmn7}
\title{Calculate the PMN given the pedo transfer function of Haer & Benbi (2003)}
\usage{
sptf_pmn7(A_C_OF, A_CLAY_MI, t = 7)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{t}{(numeric) Length of incubation period (days).}
}
\description{
Calculate the PMN given the pedo transfer function of Haer & Benbi (2003)
}
\references{
Haer and Benbi (2003) Modeling Nitrogen Mineralization Kinetics in Arable Soils of Semiarid India. Arid Land Research Management, 17, 153-168
}
