% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmn.R
\name{sptf_pmn9}
\alias{sptf_pmn9}
\title{Calculate the PMN given the pedo transfer function of Rasiah (1995), based on dataset of Campbel et al (1984)}
\usage{
sptf_pmn9(A_N_RT, CULT = 1, t = 7)
}
\arguments{
\item{A_N_RT}{(numeric) Total nitrogen content (mg N/ kg)}

\item{CULT}{(integer) whether cultivated (0) or virgin soil (0)}

\item{t}{(numeric) Length of incubation period (days).}
}
\description{
Calculate the PMN given the pedo transfer function of Rasiah (1995), based on dataset of Campbel et al (1984)
}
\references{
Campbel et al (1984)  Mineralization rate constants and their use for estimating nitrogen mineralization in some Canadian prairie soils. Canadian J. Soil Sci. 64:333-343.
}
