% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sss.R
\name{sptf_sss1}
\alias{sptf_sss1}
\title{Calculate the soil shear strength.}
\usage{
sptf_sss1(A_SOM_LOI, A_CACO3_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CACO3_MI}{(numeric) The calcium carbonate content of the soil (\%)}
}
\description{
This function calculates the soil shear strength via various soil properties
}
\references{
Zhang et al. (2018) Estimation of surface shear strength of undisturbed soils in the eastern part of northern China’s wind erosion area
}
