% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_erosion.R
\name{sptf_sss2}
\alias{sptf_sss2}
\title{Calculate the soil shear strength.}
\usage{
sptf_sss2(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
This function calculates the soil shear strength via various soil properties for soils (8-12 cm) in Denmark and Belgium
}
\references{
Schjonning (2021) Topsoil shear strength – Measurements and predictions
}
