% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sss.R
\name{sptf_sss3}
\alias{sptf_sss3}
\title{Calculate the soil shear strength.}
\usage{
sptf_sss3(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
This function calculates the soil shear strength via various soil properties for agricultural soils in Denmark
}
\references{
Schjonning (2021) Topsoil shear strength – Measurements and predictions. Using data from Schjonning et al. (2020).
}
