% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptf_textureclass.R
\name{sptf_textureclass}
\alias{sptf_textureclass}
\title{Estimate soil texture class B_TEXTURE_USDA}
\usage{
sptf_textureclass(A_CLAY_MI, A_SILT_MI, A_SAND_MI)
}
\arguments{
\item{A_CLAY_MI}{(numeric) Clay content (\%)}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_SAND_MI}{(numeric) Silt content (\%)}
}
\value{
Texture according to the USDA classification
}
\description{
Estimate soil texture class B_TEXTURE_USDA
}
