% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv.R
\name{sptf_tv1}
\alias{sptf_tv1}
\title{Predicting threshold velocity (m / sec)}
\usage{
sptf_tv1(A_SOM_LOI, A_CLAY_MI, A_AMWD_DM = NA_real_)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_AMWD_DM}{(numeric) The aggregate mean weight diameter in mm (dry aggregate stability).}
}
\description{
Calculate the threshold velocity for wind erosion for soils in Iran.
}
\references{
Rezaei et al. (2022) Determination of the threshold velocity of soil wind erosion using a wind tunnel and its prediction for calcareous soils of Iran
}
