% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv.R
\name{sptf_tv2}
\alias{sptf_tv2}
\title{Predicting threshold velocity (m / sec)}
\usage{
sptf_tv2(A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}
}
\description{
Calculate the threshold velocity for soils sensitive to wind erosion in Hungary
}
\references{
unknown
}
