% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv.R
\name{sptf_tv3}
\alias{sptf_tv3}
\title{Predicting threshold velocity (m / sec)}
\usage{
sptf_tv3(A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}
}
\description{
Calculate the threshold velocity for wind erosion for agicultural soils in Hungary
}
\references{
Negyesi et al. (2016) Effect of soil parameters on the threshold wind velocity and maximum eroded mass in a dry environment
}
