% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whc.R
\name{sptf_whc1}
\alias{sptf_whc1}
\title{Calculate the WHC given the pedotransferfunction of Bagdal et al 2022 for non-calcareous soil}
\usage{
sptf_whc1(A_C_OF, A_SAND_MI, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
Calculate the WHC given the pedotransferfunction of Bagdal et al 2022 for non-calcareous soil
}
\references{
Bagdal et al (2022) Carbon-sensitive pedotransfer functions for plant available water
}
