% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whc.R
\name{sptf_whc10}
\alias{sptf_whc10}
\title{Calculate the WHC given the pedotransferfunction of Rawls et al 1982}
\usage{
sptf_whc10(A_CLAY_MI, A_SILT_MI, A_SAND_MI, A_C_OF, D_BDS, mp_fc = 33)
}
\arguments{
\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}

\item{A_C_OF}{(numeric) The soil organic carbon content (\%).}

\item{D_BDS}{(numeric) The soil bulk density (g/cm3).}

\item{mp_fc}{(numeric) Water potential at field capacity (kPa). 10 or 33}
}
\description{
Calculate the WHC given the pedotransferfunction of Rawls et al 1982
}
\references{
Rawls et al (1982) Estimation of soil water properties. Trans. ASAE 25, 1316–1320.
}
