% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whc.R
\name{sptf_whc12}
\alias{sptf_whc12}
\title{Calculate the WHC given the pedotransferfunction of Rawls & Brakensiek 1985
Note: Calculation of wrc parameters may be wrong. The calculated water content is out of normal range. Original literature can not be found on web.}
\usage{
sptf_whc12(A_CLAY_MI, A_SAND_MI, D_BDS, mp_wp = 1500, mp_fc = 33)
}
\arguments{
\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}

\item{D_BDS}{(numeric) The soil bulk density (g/cm3).}

\item{mp_wp}{(numeric) Water potential at wilting point (kPa).}

\item{mp_fc}{(numeric) Water potential at field capacity (kPa). 10 or 33}
}
\description{
Calculate the WHC given the pedotransferfunction of Rawls & Brakensiek 1985
Note: Calculation of wrc parameters may be wrong. The calculated water content is out of normal range. Original literature can not be found on web.
}
\references{
Rawls & Brakensiek (1985) Prediction of Soil Water Properties for Hydrologic Modeling
}
