% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whc.R
\name{sptf_whc4}
\alias{sptf_whc4}
\title{Calculate the WHC given the pedotransferfunction of Oosterveld and Chang (1980)}
\usage{
sptf_whc4(A_SAND_MI, A_CLAY_MI, D_BDS, A_DEPTH = 15, mp_wp = 1500)
}
\arguments{
\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{D_BDS}{(numeric) The soil bulk density (g/cm3).}

\item{A_DEPTH}{(numeric) The mean depth of soil sample (cm). This should range within 8 - 180 cm.}

\item{mp_wp}{(numeric) Water potential at wilting point (kPa).}
}
\description{
Calculate the WHC given the pedotransferfunction of Oosterveld and Chang (1980)
}
\references{
Oosterveld and Chang (1980) Empirical relations between laboratory determinations of soil texture and moisture retention
}
