% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whc.R
\name{sptf_whc5}
\alias{sptf_whc5}
\title{Calculate the waterholding capacity given the pedotransferfunction of Wösten et al. 1999}
\usage{
sptf_whc5(A_SILT_MI, A_CLAY_MI, D_BDS, A_SOM_LOI, topsoil = 1, mp_fc = 33)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{topsoil}{(boolean) Whether top soil (1) or sub-soil (0)}

\item{mp_fc}{(numeric) Water potential at field capacity (kPa).}
}
\description{
Calculate the waterholding capacity given the pedotransferfunction of Wösten et al. 1999
}
\references{
Wösten et al. (1999) Development and use of a database of hydraulic properties of European soils
}
