% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whc.R
\name{sptf_whc9}
\alias{sptf_whc9}
\title{Calculate the WHC given the pedotransferfunction of Tomasella & Hodnett 1998}
\usage{
sptf_whc9(A_CLAY_MI, A_SILT_MI, A_C_OF, mp_fc = 33)
}
\arguments{
\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}

\item{A_C_OF}{(numeric) The soil organic carbon content (\%).}

\item{mp_fc}{(numeric) Water potential at field capacity (kPa). 10 or 33}
}
\description{
Calculate the WHC given the pedotransferfunction of Tomasella & Hodnett 1998
}
\references{
Tomasella & Hodnett (1998) Estimating soil water retention characteristics from limited data in Brazilian Amazonia. Soil Sci. 163, 190-202.
}
