% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsa.R
\name{sptf_wsa1}
\alias{sptf_wsa1}
\title{Calculate the percentage water stable aggregate}
\usage{
sptf_wsa1(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}
}
\description{
Calculate the percentage water stable aggregates for topsoils (0-20cm) in the United Kingdom given the pedotransferfunction of Stengel et al. (1984).
}
\references{
Stengel et al. (1984).Factors influencing the variation of some properties of soils in relation to their suitability for direct drilling.
}
