% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsa.R
\name{sptf_wsa10}
\alias{sptf_wsa10}
\title{Predicting the percentage water stable aggregates}
\usage{
sptf_wsa10(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}
}
\description{
Calculate the Percentage Water Stable Aggregates for a soils (0-5 cm) in Australia
}
\references{
Chan et al. (1992) Organic Carbon and Associated Soil Properties of a Red Earth after 10 years of Rotation under Different Stubble and Tillage Practices
}
