% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsa.R
\name{sptf_wsa2}
\alias{sptf_wsa2}
\title{Calculate the percentage water stable aggregate}
\usage{
sptf_wsa2(A_SOM_LOI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}
}
\description{
Calculate the percentage water stable aggregates for agricultural soils (0-15 cm) in UK given the pedotransferfunction of Ekwue (1990)
}
\references{
Ekwue (1990). Organic matter effects on soil strength properties.
}
