% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsa.R
\name{sptf_wsa3}
\alias{sptf_wsa3}
\title{Calculate the percentage water stable aggregate}
\usage{
sptf_wsa3(A_SOM_LOI, A_CLAY_MI, A_SILT_MI, A_K_AA, A_PH_WA)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_K_AA}{(numeric) Potassium plant available, extracted with acetic acid (mg K/kg)}

\item{A_PH_WA}{(numeric) Acidity in water (-)}
}
\description{
Calculate the Percentage Water Stable Aggregates for agricultural soils (0-20cm) in Turkey given the pedotransferfunction of Gulser (2018)
}
\references{
Gulser (2018) Predicting Aggregate Stability of Cultivated Soils
}
