% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsa.R
\name{sptf_wsa4}
\alias{sptf_wsa4}
\title{Predicting the percentage water stable aggregates}
\usage{
sptf_wsa4(A_CLAY_MI, A_SILT_MI, A_C_OF)
}
\arguments{
\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}

\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}
}
\description{
Calculate the percentage water stable aggregates (%)
}
\details{
This function calculates the erodible fraction in Argentina for agricultural topsoils being cultivated and uncultivated.
}
\references{
Colazo & Buschiazzo (2010) Soil dry aggregate stability and wind erodible fraction in a semiarid environment of Argentina
}
