% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterstableaggregates.R
\name{sptf_wsa5}
\alias{sptf_wsa5}
\title{Predicting the percentage water stable aggregates}
\usage{
sptf_wsa5(A_CLAY_MI, A_SILT_MI, A_SOM_LOI)
}
\arguments{
\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}

\item{A_SOM_LOI}{(numeric) The soil organic matter content (\%).}
}
\description{
Calculate the percentage water stable aggregates (%) for variable alluvial soils (0-10cm) in southern Ohio.
}
\details{
This function calculates the percentage water stable aggregates for alluvial soils in southern Ohio
}
\references{
Salchow et al. (1996) Pedotransfer functions for variable alluvial soils in southern Ohio.
}
