% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsa.R
\name{sptf_wsa6}
\alias{sptf_wsa6}
\title{Predicting the percentage water stable aggregates}
\usage{
sptf_wsa6(A_C_OF, A_CLAY_MI, A_SILT_MI, A_CACO3_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_CACO3_MI}{(numeric) The calcium carbonate content of the soil (\%)}
}
\description{
Calculate the water stable aggragates for brown calcareous soils (0-5 cm) within the Mediterranean basin in France given the pedotransferfunction of le Bissonnais et al. (2007)
}
\references{
le Bissonnais et al. (2007) Erodibility of Mediterranean vineyard soils: relevant aggregate stability methods and significant soil variables
}
