% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{sptf_yield1}
\alias{sptf_yield1}
\title{Calculate the crop yield of cereals via the meta-regression of global dataset from Oldfield et al. (2019)}
\usage{
sptf_yield1(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
This function calculates the crop yield of cereals.
}
\references{
Oldfield et al. (2019).Global meta-analysis of the relationship between soil organic matter and crop yields
}
