% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texture.classes.R
\name{texture.class}
\alias{texture.class}
\alias{texture.class.character}
\alias{texture.class.data.frame}
\alias{texture.class.numeric}
\title{Sanitize texture class descriptions}
\usage{
texture.class(x, ..., stringsAsFactors = TRUE, validate = TRUE)

\method{texture.class}{character}(x, ..., stringsAsFactors = TRUE,
  validate = TRUE)

\method{texture.class}{numeric}(x, clay, silt = NULL, ...,
  stringsAsFactors = TRUE, validate = TRUE)

\method{texture.class}{data.frame}(x, ..., stringsAsFactors = TRUE,
  validate = TRUE)
}
\arguments{
\item{x}{An object to be converted to a vector of texture classes, or a
numeric vector describing sand content.}

\item{...}{Pass to/from methods}

\item{stringsAsFactors}{Pass TRUE to output a factor vector, FALSE will generate
a character vector.}

\item{validate}{Pass TRUE to ensure output is within the 12 USDA texture classes.}

\item{clay}{A numeric vector describing relative clay content}

\item{silt}{A numeric vector describing relative silt content}
}
\value{
A factor with levels Cl, ClLo, Lo, LoSa, Sa, SaCl, SaClLo,
    SaLo, Si, SiCl, SiClLo, or SiLo; or a character vector of the above
    if \code{stringsAsFactors} is FALSE.
}
\description{
This collection of functions transforms input to a vector of USDA soil classes that
can be used as input to other functions in this package. If \code{x} is a character
or factor, it is transformed by assuming "sand" is represented by "Sa", "S", or "Sand";
"silt" is represented by "Si" or "Silt"; "clay" is represented by "Cl", "C", or "Clay";
and "loam" is represented by "Lo", "L", or "Loam". The components must be separated by
a space. If input is numeric, the input order is sand, then clay, then silt. If silt
is provided, input checked to make sure sand + clay + silt == 1 or
sand + clay + silt == 100. If all values are less than or equal to 1, a total sum of 1
is assumed; if not, a total sum of 100 is assumed. \code{x} is a \code{data.frame},
columns for sand, clay, and silt are found (as described above) and passed to
\code{texture.class.numeric}.
}
\examples{
texture.class(c("silty clay loam", "sandy clay loam", "s c l", "S C L"))

}

