% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TextureClassPolygonsUSDA}
\alias{TextureClassPolygonsUSDA}
\title{Soil Texture Polygons for the 12 USDA Texture Classes}
\format{A data frame with 58 rows and 4 variables.
\describe{
  \item{TextureClass}{The name of the texture class (one of Cl, ClLo, Lo, LoSa, Sa, SaCl, SaClLo,
    SaLo, Si, SiCl, SiClLo, or SiLo; see \link{texture.class})}
  \item{clay}{The proportion of clay (between 0 and 1)}
  \item{silt}{The proportion of silt (between 0 and 1)}
  \item{sand}{The proportion of sand (between 0 and 1)}
}}
\usage{
TextureClassPolygonsUSDA
}
\description{
Soil Texture Polygons for the 12 USDA Texture Classes
}
\keyword{datasets}

