% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd0}
\alias{sptf_bd0}
\title{Documentation of all possible bulk density pedotransfer function inputs}
\usage{
sptf_bd0(
  A_CEC_CO,
  A_CLAY_MI,
  A_C_OF,
  A_CN_FR,
  A_DEPTH,
  A_H20_T105,
  A_K_AA,
  A_N_RT,
  A_PH_CC,
  A_PH_WA,
  A_PH_KCL,
  A_SAND_M50,
  A_SAND_MI,
  A_SILT_MI,
  A_SOM_LOI,
  B_ALTITUDE,
  B_ROCKS_FR,
  B_SLOPE_ASPECT,
  B_SLOPE_DEGREE,
  mp_wp,
  mp_fc,
  D_BDS,
  A_CACO3_IF,
  A_H2O_T105
)
}
\arguments{
\item{A_CEC_CO}{(numeric) Cation Exchange Capacity (mmol+/kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CN_FR}{(numeric) Carbon / Nitrogen ratio (ratio)}

\item{A_DEPTH}{(numeric) The depth of the sampled soil layer (m)}

\item{A_H20_T105}{(numeric) The volumetric moisture content of the soil (\%)}

\item{A_K_AA}{(numeric) Potassium plant available, extracted with acetic acid (mg K/kg)}

\item{A_N_RT}{(numeric) Total nitrogen content (mg N/ kg)}

\item{A_PH_CC}{(numeric) Acidity in CaCl2 (-)}

\item{A_PH_WA}{(numeric) Acidity in water (-)}

\item{A_PH_KCL}{(numeric) Acidity in KCl (-)}

\item{A_SAND_M50}{(numeric) description (µm)}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{B_ALTITUDE}{(numeric) The altitude (m)}

\item{B_ROCKS_FR}{(numeric) description}

\item{B_SLOPE_ASPECT}{(numeric) The slope aspect (degrees)}

\item{B_SLOPE_DEGREE}{(numeric) Slope of the field (degrees)}

\item{mp_wp}{(numeric) Water potential at wilting point (kPa).}

\item{mp_fc}{(numeric) Water potential at field capacity (kPa).}

\item{D_BDS}{(numeric) Bulk density (kg/m3)}

\item{A_CACO3_IF}{(numeric) Carbonated Lime inorganic soil fraction (\%)}

\item{A_H2O_T105}{(numeric) water content measured after drying at 105 degrees (\%)}

\item{A_P_AL}{(numeric) Total phosphate content (mg P2O5/100 g)}
}
\description{
Documentation of all possible bulk density pedotransfer function inputs
}
