% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd115}
\alias{sptf_bd115}
\title{Calculate the bulk density given the pedotransferfunction of Hallet et al. (1998)}
\usage{
sptf_bd115(A_C_OF)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Hallet et al. (1998)
}
\references{
Hallet et al. (1998) Derivation and evaluation of a set of pedogenically-based empirical algorithms for predicting bulk density in British soils
}
