% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd129}
\alias{sptf_bd129}
\title{Calculate the bulk density given the pedotransferfunction of Foldal et al. (2020).}
\usage{
sptf_bd129(A_C_OF, A_CLAY_MI, A_SILT_MI, B_DEPTH)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}

\item{B_DEPTH}{(numeric) The depth of the sampled soil layer (m)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Foldal et al. (2020).
}
\references{
Foldal et al. (2020). Deriving regional pedotransfer functions to estimate soil bulk density in Austria
}
