% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd141}
\alias{sptf_bd141}
\title{Calculate the bulk density given the pedotransferfunction of Mello (2007).}
\usage{
sptf_bd141(A_C_OF, A_N_RT, A_CLAY_MI, A_SAND_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The organic C content of the soil (g / kg).}

\item{A_N_RT}{(numeric) The total N content of the soil (mg / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Mello (2007).
}
\references{
Mello (2007). Estimativas dos estoques de carbono dos solos nos Estados de Rondonia e Mato Grosso anteriores a inervencao antropica.
}
