% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd148}
\alias{sptf_bd148}
\title{Calculate the bulk density given the pedotransferfunction of Marcolin & Klein (2011).}
\usage{
sptf_bd148(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Marcolin & Klein (2011).
}
\references{
Marcolin & Klein (2011). Determinação da densidade relativa do solo por uma função de pedotransferência para a densidade do solo máxima
}
