% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd154}
\alias{sptf_bd154}
\title{Calculate the bulk density given the pedotransferfunction of Gomes et al. (2017).}
\usage{
sptf_bd154(A_C_OF, A_PH_WA)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}

\item{A_PH_WA}{(numeric) The acidity of the soil, pH in water (-)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Gomes et al. (2017).
}
\references{
Gomes et al. (2017). The use of Pedotransfer functions and the estimation of carbon stock in the Central Amazon region
}
