% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd156}
\alias{sptf_bd156}
\title{Calculate the bulk density given the pedotransferfunction of Perreault et al. (2022)}
\usage{
sptf_bd156(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Perreault et al. (2022)
}
\references{
Perreault et al. (2022) Development of Pedotransfer Functions to Predict Soil Physical Properties in Southern Quebec (Canada)
}
