% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd179}
\alias{sptf_bd179}
\title{Calculate the bulk density given the pedotransferfunction of Melendez (2017).}
\usage{
sptf_bd179(
  A_SOM_LOI,
  A_C_OF = NA,
  A_CLAY_MI = NA,
  A_SILT_MI = NA,
  A_SAND_MI = NA,
  A_PH_WA = NA
)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_PH_WA}{(numeric) Acidity in water (-)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Melendez (2017).
}
\references{
Melendez (2017) PREDICTING SOIL BULK DENSITY USING HIERARCHICAL PEDOTRANSFER FUNCTIONS AND ARTIFICIAL NEURAL NETWORKS
}
