% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd181}
\alias{sptf_bd181}
\title{Calculate the bulk density given the pedotransferfunction of Abdel (2019).}
\usage{
sptf_bd181(A_SOM_LOI, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage of organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Abdel (2019).
}
\references{
Abdel (2019). Linear Regression Models to EstimatevExchangeable Sodium Percentage and Bulk Density of Salt Affected Soils in Sahl El-Hossinia, El-Sharkia Governorate, Egypt
}
