% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd185}
\alias{sptf_bd185}
\title{Calculate the bulk density given the pedotransferfunction of Schjonning et al. (2016)}
\usage{
sptf_bd185(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Schjonning et al. (2016)
}
\references{
Schjonning et al. (2016) Predicting soil particle density from clay and soil organic matter contents.
}
