% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd79}
\alias{sptf_bd79}
\title{Calculate the bulk density given the pedotransferfunction of Brahim et al. (2012)}
\usage{
sptf_bd79(A_C_OF, A_SILT_MI, A_SAND_MI, A_PH_WA)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%)}

\item{A_PH_WA}{(numeric) The acidity of the soil, pH in water (-)}
}
\description{
Calculate the bulk density given the pedotransferfunction of Brahim et al. (2012)
}
\references{
Brahim et al. (2012). Pedotransfer functions to estimate soil bulk density for Northern Africa: Tunisia case
}
