% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd87}
\alias{sptf_bd87}
\title{Calculate the bulk density given the pedotransferfunction of Minasny & Hartemink (2011)}
\usage{
sptf_bd87(A_SOM_LOI, A_SAND_MI, A_DEPTH)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The organic matter content of the soil (\%).}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}

\item{A_DEPTH}{(numeric) The depth of the sampled soil layer (m).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Minasny & Hartemink (2011)
}
\references{
Minasny & Hartemink (2011) Predicting soil properties in the tropics
}
