% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd90}
\alias{sptf_bd90}
\title{Calculate the bulk density given the pedotransferfunction of Dexter (2004)}
\usage{
sptf_bd90(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Dexter (2004)
}
\references{
Dexter (2004). Soil physical quality Part I. Theory, effects of soil texture, density, and organic matter, and effects on root growth
}
