% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkdensity.R
\name{sptf_bd97}
\alias{sptf_bd97}
\title{Calculate the bulk density given the pedotransferfunction of Sevastas et al. (2018)}
\usage{
sptf_bd97(A_C_OF, A_SAND_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The fraction organic carbon in the soil (g / kg).}

\item{A_SAND_MI}{(numeric) The sand content of the soil (\%).}
}
\description{
Calculate the bulk density given the pedotransferfunction of Sevastas et al. (2018)
}
\references{
Sevastas et al. (2018). Predicting bulk density using pedotransfer functions for soils in the Upper Anthemountas basin, Greece
}
