% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptf_tables.R
\docType{data}
\name{sptf_bouwsteen}
\alias{sptf_bouwsteen}
\title{Table with supprting infromation for class ptfs pF curve}
\format{
A data.frame with x rows and x columns:
\describe{
  \item{bouwsteen}{}
  \item{omschrijving}{}
  \item{thres}{}
  \item{thsat}{}
  \item{Ks}{}
  \item{alpha}{}
  \item{l}{}
  \item{n}{}
  \item{sand}{}
  \item{silt}{}
  \item{clay}{}
  \item{OM}{}
  \item{bulkdensity}{}
  \item{M50}{}
  }
}
\usage{
sptf_bouwsteen
}
\description{
This table contains parameters for Dutch categorial ptfs for pF curve
}
\keyword{datasets}
