% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdec.R
\name{sptf_cdec2}
\alias{sptf_cdec2}
\title{Calculate the carbon decomposition for mineral soils using MINIP approach}
\usage{
sptf_cdec2(A_C_OF, years)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{years}{(numeric) The number of years for which the carbon decomposition need to be estimated.}
}
\description{
This function calculates the decomposition of sand and peat soils using first order kinetics.
}
\references{
Janssen (1984) A simple method for calculating decomposition and accumulation of 'young' soil organic matter.
}
