% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec34}
\alias{sptf_cec34}
\title{Calculate the CEC}
\usage{
sptf_cec34(A_C_OF, A_CLAY_MI, A_PH_CC, B_LU_PTFCLASS)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_PH_CC}{(numeric) The acidity of the soil, pH in CaCl2 (-)}

\item{B_LU_PTFCLASS}{(character) The land use category (options: agriculture, grassland, cropland, forest, nature)}
}
\description{
This function calculates the CEC at pH 7 for forest and natural grassland soils in eastern Canada
}
\references{
Meyer et al. (1994).  Cation exchange capacities of upland soils in eastern Canada.
}
