% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec46}
\alias{sptf_cec46}
\title{Calculate the CEC}
\usage{
sptf_cec46(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
This function calculates the CEC at pH 7 at agricultural sites at various depths in Tanzania
}
\references{
Mtama et al. (2018) Pedotransfer Functions for Cation Exchange Capacity, Available Water Holding Capacity and Soil Organic Carbon for Representative Soils of Southern Highland Zone of Tanzania
}
