% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec50}
\alias{sptf_cec50}
\title{Calculate the CEC}
\usage{
sptf_cec50(A_SOM_LOI, A_CLAY_MI, A_SILT_MI, A_CACO3_IF, A_PH_CC)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_CACO3_IF}{(numeric) Carbonated Lime inorganic soil fraction (\%)}

\item{A_PH_CC}{(numeric) Acidity in CaCl2 (-)}
}
\description{
This function calculates the CEC at pH 8.2 for calcareous soils in Iran, Texas, USA, Cuba and Kenya
}
\references{
Razzaghi et al. (2021) Evaluating models to estimate cation exchange capacity of calcareous soils
}
