% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec59}
\alias{sptf_cec59}
\title{Calculate the CEC}
\usage{
sptf_cec59(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
This function calculates the CEC at pH 8.2 for soil in Lower Namoi Valley, Australia.
}
\references{
McBratney et al. (2002). From pedotransfer functions to soil inference systems.
}
