% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec62}
\alias{sptf_cec62}
\title{Calculate the CEC}
\usage{
sptf_cec62(A_C_OF, A_CLAY_MI, A_SILT_MI, A_CACO3_IF)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_CACO3_IF}{(numeric) Carbonated Lime inorganic soil fraction (\%)}
}
\description{
This function calculates the CEC at pH 8.2 for calcareous soils in Iran
}
\references{
Kashi et al. (2014) Estimation of Soil Infiltration and Cation Exchange Capacity Based on Multiple Regression, ANN (RBF, MLP), and ANFIS Models.
}
