% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{sptf_cec63}
\alias{sptf_cec63}
\title{Calculate the CEC}
\usage{
sptf_cec63(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
This function calculates the CEC for soils in Italy
}
\references{
Bazzoffi et al. (1994) Statistical models for predicting aggregate stability from intrinsic soil components.
}
