% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptf_tables.R
\docType{data}
\name{sptf_countries}
\alias{sptf_countries}
\title{Table with country codes used in the package}
\format{
A data.frame with x rows and x columns:
\describe{
  \item{country_name}{the name of the country}
  \item{country_code}{the country code;from https://datahub.io/core/country-codes}
}
}
\usage{
sptf_countries
}
\description{
This table contains country names as used for the spatial extent where sptfs have been derived from.
}
\keyword{datasets}
