% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwc.R
\name{sptf_hwc5}
\alias{sptf_hwc5}
\title{Calculate the HWC content}
\usage{
sptf_hwc5(A_C_OF, A_P_AL, A_PH_CC)
}
\arguments{
\item{A_C_OF}{(numeric) The carbon content of the soil (g / kg).}

\item{A_P_AL}{(numeric) The phosphorus content of the soil, extracted in ammonium lactate (mg P2O5/100g)}

\item{A_PH_CC}{(numeric) The acidity of the soil, pH in CaCl2 (-)}
}
\description{
This function calculates the hot water extractable carbon content for agricultural soils in the Netherlands. Experimental data from 2019.
}
\references{
Van Balen et al. (2015) Effecten bodem- en structuurverbeteraars. Onderzoek op klei- en zandgrond 2010-2015. PPO-report 693.
}
