% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwd.R
\name{sptf_mwd11}
\alias{sptf_mwd11}
\title{Calculate the Mean Weight Diameter}
\usage{
sptf_mwd11(A_C_OF, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Calculate the Mean Weight Diameter given for soils (0-20cm) in Italy the pedotransferfunction of Bazzoffi et al. (1994)
}
\references{
Bazzoffi et al. (1994) Statistical models for predicting aggregate stability from intrinsic soil components.
}
