% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterstableaggregates.R
\name{sptf_mwd14}
\alias{sptf_mwd14}
\title{Predicting the the Mean Weight Diameter for soils}
\usage{
sptf_mwd14(A_SOM_LOI, A_CLAY_MI, A_SILT_MI, A_PH_WA, A_CACO3_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}

\item{A_PH_WA}{(numeric) The acidity of the soil, pH in water (-)}

\item{A_CACO3_MI}{(numeric) The calcium carbonate content of the soil (\%)}
}
\description{
Calculate the Mean Weight Diamater for agricultural soils (0-10cm) in southern Spain given the pedotransferfunction of Canasveras et al. (2010)
}
\references{
Canasveras et al. (2010) Estimation of aggregate stability indices in Mediterranean soils by diffuse reflectance spectroscopy
}
