% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{sptf_ph2}
\alias{sptf_ph2}
\title{Calculate the pH-water value from pH-KCL}
\usage{
sptf_ph2(A_SOM_LOI, A_PH_KCL, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_PH_KCL}{(numeric) Acidity in KCl (-)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
This function calculates the pH extracted with water from the pH-KCL.
}
\references{
Wosten et al. (1997) Bodemkundige vertaalfuncties bij SC-DLO; state of the art. https://edepot.wur.nl/299664
}
