% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmn.R
\name{sptf_pmn14}
\alias{sptf_pmn14}
\title{Calculate the Potentially Mineralizable N (PMN) in mg/kg}
\usage{
sptf_pmn14(A_C_OF, A_P_AL, A_PH_CC)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_P_AL}{(numeric) Total phosphate content (mg P2O5/100 g)}

\item{A_PH_CC}{(numeric) Acidity in CaCl2 (-)}
}
\description{
This function calculates the PMN content for agricultural soils in the Netherlands. PMN refers to anearobic mineralizable N pool, 40 degrees for 7 days.
Experimental data from 2019.
}
\references{
Van Balen et al. (2015) Effecten bodem- en structuurverbeteraars. Onderzoek op klei- en zandgrond 2010-2015. PPO-report 693.
}
