% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmn.R
\name{sptf_pmn8}
\alias{sptf_pmn8}
\title{Calculate the PMN given the pedo transfer function of Rasiah (1995), based on dataset of Herlihy (1979)}
\usage{
sptf_pmn8(A_N_RT, A_CEC_CO, A_SILT_MI, t = 7)
}
\arguments{
\item{A_N_RT}{(numeric) Total nitrogen content (mg N/ kg)}

\item{A_CEC_CO}{(numeric) Cation Exchange Capacity (mmol+/kg)}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{t}{(numeric) Length of incubation period (days).}
}
\description{
Calculate the PMN given the pedo transfer function of Rasiah (1995), based on dataset of Herlihy (1979)
}
\references{
Herlihy (1979) Nitrogen mineralization in soils of varying texture, moisture and organic matter. Plant and Soil. 53:255-267.
}
