% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whc.R
\name{sptf_whc1}
\alias{sptf_whc1}
\title{Calculate the water holding capacity given the pedotransferfunction of Bagdal et al.2022 for non-calcareous soil}
\usage{
sptf_whc1(A_C_OF, A_SAND_MI, A_CLAY_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}
}
\description{
Calculate the water holding capacity given the pedotransferfunction of Bagdal et al.2022 for non-calcareous soil
}
\references{
Bagdal et al.(2022) Carbon-sensitive pedotransfer functions for plant available water
}
