% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whc.R
\name{sptf_whc10}
\alias{sptf_whc10}
\title{Calculate the water holding capacity given the pedotransferfunction of Rawls et al.1982}
\usage{
sptf_whc10(A_CLAY_MI, A_SILT_MI, A_SAND_MI, A_C_OF, D_BDS, mp_fc = 33)
}
\arguments{
\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{mp_fc}{(numeric) Water potential at field capacity (kPa).}
}
\description{
Calculate the water holding capacity given the pedotransferfunction of Rawls et al.1982
}
\details{
mp_fc must be either 10 or 33
}
\references{
Rawls et al.(1982) Estimation of soil water properties. Trans. ASAE 25, 1316–1320.
}
