% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whc.R
\name{sptf_whc12}
\alias{sptf_whc12}
\title{Calculate the water holding capacity given the pedotransferfunction of Rawls & Brakensiek 1985}
\usage{
sptf_whc12(A_C_OF, A_CLAY_MI, A_SAND_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SAND_MI}{(numeric) Sand content (\%)}
}
\description{
DEPRECATED
Note: Calculation of whc parameters may be wrong. The calculated water content is out of normal range. Original literature cannot be found on web.
}
\details{
mp_fc must be either 10 or 33
}
\references{
Rawls & Brakensiek (1985) Prediction of Soil Water Properties for Hydrologic Modelling
}
