% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whc.R
\name{sptf_whc13}
\alias{sptf_whc13}
\title{Calculate the water holding capacity given the pedotransferfunction of Tian et al. 2021}
\usage{
sptf_whc13(A_C_OF, A_CLAY_MI, A_SAND_MI)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SAND_MI}{(numeric) Sand content (\%)}
}
\description{
Calculate the water holding capacity given the pedotransferfunction of Tian et al. 2021
}
\references{
Tian et al. (2021) New pedotransfer functions for soil water retention curves that better account for bulk density effects
}
