% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whc.R
\name{sptf_whc16}
\alias{sptf_whc16}
\title{the water holding capacity given the pedotransferfunction of Wösten et al.2001 (Table 3), for each soil class}
\usage{
sptf_whc16(A_SOM_LOI, A_CLAY_MI, A_SILT_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil determined by loss on ignition (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_SILT_MI}{(numeric) Silt content (\%)}

\item{mp_wp}{(numeric) Water potential at wilting point (kPa).}
}
\description{
the water holding capacity given the pedotransferfunction of Wösten et al.2001 (Table 3), for each soil class
}
\references{
Wösten, J. H. M., Veerman, G. ., de Groot, W. J., & Stolte, J. (2001). Waterretentie en doorlatendheidskarakteristieken van boven- en ondergronden in Nederland: de Staringreeks. Alterra Rapport, 153, 86. https://edepot.wur.nl/43272
}
