% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whc.R
\name{sptf_whc4}
\alias{sptf_whc4}
\title{Calculate the waterholding capacity given the pedotransferfunction of Oosterveld and Chang (1980)}
\usage{
sptf_whc4(A_SAND_MI, A_CLAY_MI, D_BDS, A_DEPTH = 15, mp_wp = 1500)
}
\arguments{
\item{A_SAND_MI}{(numeric) Sand content (\%)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_DEPTH}{(numeric) The depth of the sampled soil layer (m)}

\item{mp_wp}{(numeric) Water potential at wilting point (kPa).}
}
\description{
Calculate the waterholding capacity given the pedotransferfunction of Oosterveld and Chang (1980)
}
\details{
For this function A_DEPTH should be between 0.08 and 0.18
}
\references{
Oosterveld and Chang (1980) Empirical relations between laboratory determinations of soil texture and moisture retention
}
