% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whc.R
\name{sptf_whc6}
\alias{sptf_whc6}
\title{Calculate the water holding capacity given the pedotransferfunction of Vereecken et al. 1989}
\usage{
sptf_whc6(A_SAND_MI, A_CLAY_MI, D_BDS, A_C_OF, mp_fc = 33)
}
\arguments{
\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{A_CLAY_MI}{(numeric) The clay content of the mineral soil fraction (\%).}

\item{A_C_OF}{(numeric) Organic Carbon Content (g / kg)}

\item{mp_fc}{(numeric) Water potential at field capacity (kPa).}
}
\description{
Calculate the water holding capacity given the pedotransferfunction of Vereecken et al. 1989
}
\references{
Vereecken et al. (1989) Estimating the soil moisture retention characteristics from texture, bulk density and carbon content
}
