% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsa.R
\name{sptf_wsa3}
\alias{sptf_wsa3}
\title{Calculate the percentage water stable aggregate}
\usage{
sptf_wsa3(A_SOM_LOI, A_CLAY_MI, A_SILT_MI, A_K_AA, A_PH_WA)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_SILT_MI}{(numeric) The silt content of the soil (\%).}

\item{A_K_AA}{(numeric) The potassium level in soil, extrated via ammonium acetate (mg/kg)}

\item{A_PH_WA}{(numeric) The acidity of the soil, pH in water (-)}
}
\description{
Calculate the Percentage Water Stable Aggregates for agricultural soils (0-20cm) in Turkey given the pedotransferfunction of Gulser (2018)
}
\references{
Gulser (2018) Predicting Aggregate Stability of Cultivated Soils
}
