% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsa.R
\name{sptf_wsa8}
\alias{sptf_wsa8}
\title{Predicting the percentage water stable aggregates}
\usage{
sptf_wsa8(A_SOM_LOI, A_CLAY_MI)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}
}
\description{
Calculate the Percentage Water Stable Aggregates for tiled soils in Ontaria, Canada, given the pedotransferfunction of Perfect et al. (1993)
}
\references{
Perfect et al. (1993) Comparison of functions for characterizing the dry aggregate size distribution of tilled soil
}
