% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterstableaggregates.R
\name{sptf_wsa9}
\alias{sptf_wsa9}
\title{Predicting the percentage water stable aggregates}
\usage{
sptf_wsa9(A_SOM_LOI, A_CLAY_MI, A_PH_WA)
}
\arguments{
\item{A_SOM_LOI}{(numeric) The percentage organic matter in the soil (\%).}

\item{A_CLAY_MI}{(numeric) The clay content of the soil (\%).}

\item{A_PH_WA}{(numeric) The acidity of the soil, pH in water (-)}
}
\description{
Calculate the Percentage Water Stable Aggregates for agricultural and forest soils (0-17cm) in Chile given the pedotransferfunction of Rivera & Bonilla (2020)
}
\references{
Rivera & Bonilla (2020) Predicting soil aggregate stability using readily available soil properties and machine learning techniques
}
